% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{adapt_Morphoscape}
\alias{adapt_Morphoscape}
\title{Adapt format of Morphoscape objects}
\usage{
adapt_Morphoscape(obj)
}
\arguments{
\item{obj}{An object of class \code{"kriged_surfaces"},
\code{"wtd_lscp"}, \code{"poly_surf"} or \code{"multi_poly"}.}
}
\value{
A list containing the x and y values of the grid, and a matrix
with the z values of each combination of x and y.
}
\description{
Extract surface information from landscapes created
with \code{Morphoscape} and adapt format to be used by
\code{\link[graphics:contour]{graphics::contour()}} and \code{\link[graphics:image]{graphics::image()}}.
}
\keyword{internal}
