% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to scatterplots}
\usage{
add_labels(xy, labels = NULL)
}
\arguments{
\item{xy}{A matrix with scatter point coordinates.}

\item{labels}{Either logical, indicating whether to add labels to all the
points in the scatter plot (taken from row names), or character string
containing specific names to be added.}
}
\description{
A wrapper for \code{\link[graphics:text]{graphics::text()}}. Used internally.
}
\keyword{internal}
