% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_operations.R
\name{ax_transformation}
\alias{ax_transformation}
\title{Compute shapes at the extremes of a morphometric axis}
\usage{
ax_transformation(obj, axis = 1, mag = 1, datype = "landm")
}
\arguments{
\item{obj}{An object containing either a multivariate ordination of class
\code{"prcomp", "bg_prcomp", "phy_prcomp"} or \code{"pls_shape"} or a
\code{"mlm"} object fitted using \code{\link[=lm]{lm()}}.}

\item{axis}{An optional numeric value specifying the axis of the multivariate
ordination which is to be represented.}

\item{mag}{Numeric; magnifying factor for representing shape transformation.}

\item{datype}{Character; the type of shape data (either \code{"fcoef"} or
\code{"landm"}).}
}
\value{
Shapes representing the extremes of the morphometric axis. For
landmark data, a \code{p x k x 2} array. For fourier data, a \code{2 x (4 x
  nb.h)} matrix (\code{nb.h} being the number of harmonics used during
elliptic Fourier analysis).
}
\description{
This function computes the theoretical shapes corresponding to the
extremes of a morphometric axis, which can be supplied as an object
containing a linear model or a multivariate ordination analysis.
}
\examples{
#load tail data
data("tails")

#perform PCA, compute and plot extreme shapes of PC1 at its natural range
pca <- prcomp(geomorph::two.d.array(tails$shapes))
extshapes <- ax_transformation(obj = pca, axis = 1, mag = 1)
pile_shapes(extshapes, links = links, mshape = FALSE)

#compute and plot extreme shapes of PC2 at its natural range
extshapes <- ax_transformation(obj = pca, axis = 2, mag = 1)
pile_shapes(extshapes, links = links, mshape = FALSE)

#compute and plot extreme shapes of PC2 magnified x2
extshapes <- ax_transformation(obj = pca, axis = 2, mag = 2)
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes at its natural range
model <- lm(geomorph::two.d.array(tails$shapes) ~ log(tails$sizes))
extshapes <- ax_transformation(obj = model, mag = 1)
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes at its natural range
model <- lm(geomorph::two.d.array(tails$shapes) ~ tails$data$sex)
extshapes <- ax_transformation(obj = model, mag = 1)
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes magnified x2
model <- lm(geomorph::two.d.array(tails$shapes) ~ tails$data$sex)
extshapes <- ax_transformation(obj = model, mag = 2)
pile_shapes(extshapes, links = links, mshape = FALSE)
}
\references{
MacLeod, N. (2009). \emph{Form & shape models}. Palaeontological
Association Newsletter, 72(620), 14-27.
}
