% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{bg_prcomp}
\alias{bg_prcomp}
\title{Between-groups Principal Component Analysis}
\usage{
bg_prcomp(
  x,
  groups,
  gweights = TRUE,
  LOOCV = FALSE,
  recompute = FALSE,
  corr = FALSE
)
}
\arguments{
\item{x}{A matrix with variables as columns and observations as rows.}

\item{groups}{Factor; classification of observations of \code{x} into a
priori groups.}

\item{gweights}{Logical; whether to weight each group by its number of
observations.}

\item{LOOCV}{Logical; whether to apply leave-one-out cross-validation.}

\item{recompute}{Logical; whether to re-compute rotation matrix using the
scores resulting from LOOCV.}

\item{corr}{Logical; whether to use correlation instead of covariance matrix
as input.}
}
\value{

}
\description{
Perform between group PCA allowing for group weighting and
leave-one-out cross-validation, which is useful one the number of variables
exceeds the number of observations (i.e., alleviates spurious separation
between groups).
}
