% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{burnaby}
\alias{burnaby}
\title{Burnaby's orthogonal subspace computation}
\usage{
burnaby(
  x,
  vars = NULL,
  tree = NULL,
  evmodel = "BM",
  axmat = NULL,
  center = NULL
)
}
\arguments{
\item{x}{A matrix with one or more variables as columns and observations as
rows, representing the variables the user desires to be rendered
independent (orthogonal) to \code{vars}.}

\item{vars}{A matrix with one or more variables as columns and observations
as rows, representing the source of variation the user wants to remove from
\code{x}.}

\item{tree}{A \code{"phylo"} object containing a phylogenetic tree. Tip
labels should match the row names from \code{x} and \code{vars}.}

\item{evmodel}{Character, specifying an evolutionary model to perform
ancestral character reconstruction; options are "BM" (Brownian motion),
"EB" (Early burst) and "lambda" (Pagel's lambda transformation) (see
\code{\link[mvMORPH]{mvgls}} for more details).}

\item{axmat}{An optional matrix of axes coefficients to render \code{x}
orthogonal to. If provided, is used instead of \code{vars} to orthogonalize
\code{x}.}

\item{center}{An optional numeric string containing the center of the
data, to be used to re-center the data if \code{axmat} has been provided.
Ignored if \code{vars} has been specified.}
}
\value{
A \code{"burnaby"} or \code{"phy_burnaby"} object, containing:
\itemize{
\item \code{$x:} a matrix with the scores of observations in the new
ordination axes.
\item \code{$sdev:} the standard deviations of the principal components
(i.e., the square roots of the eigenvalues of the covariance matrix).
\item \code{$rotation:} a matrix of eigenvector coefficients.
\item \code{$center:} the mean values of the original variables (or their
phylogenetic mean, if \code{tree} is provided).
\item \code{$totvar:} the sum of the variances of the original data.
}
}
\description{
Performs Burnaby's multivariate approach for standardizing
morphometric data relative to nuisance factors by computing and projecting
data onto an orthogonal subspace.
}
\details{
Originally devised by Burnaby (1966) to eliminate allometric
variation from morphometric data sets by computing a subspace that is
orthogonal (i.e., mathematically independent) to an allometric vector,
and then projecting the data onto these new set of axes (However, this
approach can be used to standardize morphometric data for any type of
arbitrary vector(s) -- commonly one or more nuisance factors or covariates
introducing variation the user wishes to exclude from the analysis).

In particular, this function is intended to provide a way for computation
of ordination (sub)spaces orthogonal to other specific sources of
variation, as well as visualization of shape variation associated to their
ordination axes, most commonly as part of the \code{\link{mspace}} \%>\%
\code{proj_*} pipeline. If the user wishes to use the Burnaby approach for
standardization of morphometric (shape) data, this is better achieved
through \code{\link{detrend_shapes}} by setting
\code{method = "orthogonal"}.

The subspace produced in this way has one less dimension for each axis or
variable used for standardization, relative to the original variables.
However, the user should be aware that, when analyzing shape variation, a
rather large portion of variation in \code{x} that is not orthogonal to
\code{vars} will end up 'hidden' in the lowest ordination axes (i.e., those
effectively lost during standardization procedures that remove non-shape
variation, and therefore showing null eigenvalues). Therefore, projection
of the original shape variables into the orthogonal subspace should be
attempted after removing (at least) these null axes.

The phylogenetic version of this method (i.e., the correction of
non-phylogenetic independence from the relationship between \code{x} and
\code{vars}) is achieved through \code{\link[mvMORPH]{mvgls}}.
}
\examples{
#load packages
library(geomorph)
library(Morpho)

#extract input data: shapes and sizes from S. vacaensis.
data("shells3D")
index <- shells3D$data$species == "vacaensis"
allshapes <- shells3D$shapes
shapes <- allshapes[,,index]
sizes <- log(shells3D$sizes[index])

#construct template for visualization
template <- tps3d(x = shells3D$mesh_meanspec,
                  refmat = allshapes[,,findMeanSpec(allshapes)],
                  tarmat = expected_shapes(shapes))


#find shape subspace orthogonal to size using the Burnaby approach for
#S. vacaensis
burn <- burnaby(x = geomorph::two.d.array(shapes), vars = sizes)

#inspect results
names(burn) #the contents of the resulting object
exp_var(burn) #interpretation of eigenvalues is complex! not yet implemented
plot(burn$x) #ordination


#compare shape variation as summarized by different methods

if (interactive()) {

#build morphospace for S. vacaensis using Burnaby's approach (orthogonal
#to size variation)
msp1 <- mspace(shapes, template = template, cex.ldm = 0, bg.models = "gray",
               adj_frame = c(0.9, 0.9), asp.models = 1.1, alpha.models = 0.7,
               FUN = burnaby, vars = sizes) \%>\%
  proj_shapes(shapes, pch = 16)

#compare against a morphospace made using regular PCA
mspace(shapes, template = template, cex.ldm = 0, bg.models = "gray",
       adj_frame = c(0.9, 0.9), asp.models = 1.1, alpha.models = 0.7) \%>\%
  proj_shapes(shapes, pch = 16)


#extract sample and background shape models for subspace for S. vacaensis
grid_shapes <- extract_shapes(msp1)$shapes
na_shapes <- extract_shapes(msp1, scores = msp1$projected$scores[,1:2])$shapes

#perform regression of size on shape for S. vacaensis
reg <- lm(two.d.array(shapes) ~ sizes)

#build morphospace for the entire dataset using regular PCA, project both the
#empirical shapes, orthogonal subspace range, and size-standardized shapes
mspace(allshapes, template = template, cex.ldm = 0,
       bg.models = "gray", adj_frame = c(0.9, 0.9), asp.models = 1.1,
       alpha.models = 0.7) \%>\%
  proj_shapes(shapes, pch = 1) \%>\%
  proj_shapes(na_shapes, pch = 21, bg = "red") \%>\%
  proj_axis(reg, lwd = 2) \%>\%
  proj_groups(grid_shapes, alpha = 0.2, col = "red")
}
}
\references{
Burnaby, T. P. (1966) \emph{Growth-invariant discriminant functions and
generalized distances. Biometrics, 22, 96–110.}

Claude, J. (2008). \emph{Morphometrics with R}. Springer Science & Business
Media, 316.

Revell, L. J. (2009). \emph{Size-correction and principal components for
interspecific comparative studies}. Evolution, 63, 3258-3268.

Klingenberg, C. P. (2016). \emph{Size, shape, and form: concepts of allometry
in geometric morphometrics}. Development Genes and Evolution, 226(3),
113-137.
}
\seealso{
\code{\link{detrend_shapes}}, \code{\link{exp_var}}
}
