% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_operations.R
\name{consensus}
\alias{consensus}
\title{Compute consensus shape(s)}
\usage{
consensus(shapes, index = NULL)
}
\arguments{
\item{shapes}{Shape data.}

\item{index}{Either a numeric vector indicating the configurations to be
averaged or a factor whose levels are used to average groups of
configurations.}
}
\value{
Either a matrix defining a single mean shape \code{p x k} or an array
of shapes, one for each level of \code{index}.
}
\description{
Compute the mean shape from the entire sample, the mean shape of
a subset of samples, or the mean shape of the levels of a factor (for
landmark configurations).
}
