% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_operations.R
\name{correct_efourier}
\alias{correct_efourier}
\title{Correct artificially rotated set of Fourier shapes interactively}
\usage{
correct_efourier(ef, index = NULL)
}
\arguments{
\item{ef}{An \code{"OutCoe"} object.}

\item{index}{An optional integer vector providing the indices identifying the
shapes to be rotated. If declared, overrides interactive selection of
shapes.}
}
\value{
An \code{"OutCoe"} object containing the corrected outlines.
}
\description{
Choose and correct 180-degrees spurious rotation in a sample of
closed outline shapes.
}
\details{
The usage of this function is inspired in the SHAPE program for
elliptic Fourier analysis (Iwata & Ukai 2002). The \code{index} argument is
intended to facilitate its inclusion in scripts without having to manually
select the outlines every time.
}
\examples{
#load shells data
data("shells")
ef <- shells$shapes

#correct first 3 shapes automatically with index
ef_corr1 <- correct_efourier(ef, index = 1:3)
pile_shapes(ef_corr1, mshape = FALSE)

#correct interactively (the process will remain open until
#you click 'finish' or press the Esc key)
\dontrun{

ef_corr2 <- correct_efourier(ef_corr1)
}
}
\references{
Iwata, H., & Ukai, Y. (2002). \emph{SHAPE: a computer program package for
quantitative evaluation of biological shapes based on elliptic Fourier
descriptors}. Journal of Heredity, 93(5), 384-385.
}
