% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{density_by_group_2D}
\alias{density_by_group_2D}
\title{Plot univariate density distributions for a series of groups}
\usage{
density_by_group_2D(
  xy,
  fac,
  ax,
  alpha = 0.2,
  lwd = 1,
  lty = 1,
  plot = TRUE,
  col = seq_len(nlevels(fac))
)
}
\arguments{
\item{xy}{Coordinates of the scatterplot.}

\item{fac}{A factor grouping data points.}

\item{ax}{the axis of \code{xy} corresponding to the active variable.}

\item{alpha}{Numeric; transparency factor for density distributions.}

\item{lwd}{A vector indicating the width of line (integer) used to draw
density distributions.}

\item{lty}{A vector indicating the type of line (integer) used to draw density
distributions.}

\item{col}{A vector (either character or numeric) indicating the colors used
for each group.}
}
\value{
None
}
\description{
Plot density distribution for different groups in "univariate"
scatterplots. Used internally.
}
\examples{
#load Fourier data and necessary packages
library(geomorph)
data("shells")
shapes <- shells$shapes$coe
species <- shells$data$species

#perform PCA
pca <- prcomp(shapes)

#bind 1st axis with a column of 0s, plot and add density distributions
xy <- cbind(pca$x[,1], 0)
plot(xy, ylim = c(0,1))
density_by_group_2D(xy, fac = species, ax = 1)
}
\keyword{internal}
