% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_operations.R
\name{detrend_shapes}
\alias{detrend_shapes}
\title{Remove shape variation associated to external variables}
\usage{
detrend_shapes(model, xvalue = NULL, newx = NULL, newy = NULL)
}
\arguments{
\item{model}{A \code{mlm} object created using \code{\link[=lm]{lm()}}.}

\item{xvalue}{A value (numeric) or level (character) at which shape data is
to be standardized (i.e. centered); If NULL, the mean of the complete
sample is used.}

\item{newx, newy}{New data to be standardized instead of the used in
\code{model}. Coefficients are taken from the linear model and applied to
the new data to predict shapes at the desired \code{xvalue}.}
}
\value{

}
\description{
Detrend (i.e. standardize) shape data using the functional
relationship between shape data and some external explanatory variable(s)
(works for both factors and numerics), estimated using a linear model.
}
