% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ellipses_by_group_2D}
\alias{ellipses_by_group_2D}
\title{Plot 2D confidence ellipses for a series of groups}
\usage{
ellipses_by_group_2D(
  xy,
  fac,
  col = seq_len(nlevels(fac)),
  lty = 1,
  conflev = 0.95,
  alpha = 0,
  ...
)
}
\arguments{
\item{xy}{Coordinates of the scatterplot.}

\item{fac}{A factor grouping data points.}

\item{col}{A vector (either character or numeric) indicating the colors used
for each group.}

\item{lty}{A numeric vector indicating the type of line used to draw ellipses.}

\item{conflev}{Numeric, specifying the confidence level for drawing ellipses.}

\item{alpha}{Numeric; transparency factor for ellipses.}

\item{...}{Further arguments passed to \code{\link[graphics:polygon]{graphics::polygon()}}.}
}
\description{
Plot confidence ellipses for different groups in 2D scatterplots
created using the generic \code{\link[graphics:plot.default]{graphics::plot()}} function. Used internally (mostly).
}
\examples{
#load landmark data and necessary packages
library(geomorph)
data("wings")
shapes <- wings$shapes
species <- wings$data$species

#perform PCA
pca <- prcomp(two.d.array(shapes))

#plot and add 95\% confidence ellipses
plot(pca$x)
ellipses_by_group_2D(pca$x, fac = species, col = "black", conflev = 0.95)
}
\seealso{
\code{\link{hulls_by_group_2D}}, \code{\link[car]{ellipse}}
}
