% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{exp_var}
\alias{exp_var}
\title{Calculate percentages of variation accounted by synthetic axes}
\usage{
exp_var(ordination)
}
\arguments{
\item{ordination}{An ordination (i.e. a \code{"prcomp"}, \code{"bg_prcomp"},
\code{"phy_prcomp"}, \code{"pls"} or \code{"phy_pls"} object).}
}
\value{
A table informing the percentages and cumulative percentages of
original variation accounted by each synthetic axis of the multivariate
ordination.
}
\description{
Calculate the percentage of total original variation accounted
by syntethtic axes generated by different multivariate ordination methods.
}
\seealso{
\code{\link[base]{prcomp}}, \code{\link{bg_prcomp}},
\code{\link{phy_prcomp}}, \code{\link{pls_shapes}}
}
