% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_operations.R
\name{expected_shapes}
\alias{expected_shapes}
\title{Compute shapes expected by a linear model}
\usage{
expected_shapes(model, xvalue)
}
\arguments{
\item{model}{A \code{mlm} object created using \code{\link[=lm]{lm()}}.}

\item{xvalue}{A numeric value or values of \code{x} in the \code{model} at
which calculate expected shape(s).}
}
\value{
A 2-margins matrix, of dimensions \code{n x (k x p)} for the case of
landmark data (where \code{n} is the number of shapes provided in the left
side of the formula in \code{model}) and \code{n x (4 x nb.h)} for the case
of Fourer data (where nb. h is the number of harmonics used in elliptic
Fourier analysis).
}
\description{
This function calculates and returns shapes expected under a
linear model, at one or more specific values of a numeric explanatory
variable.
}
\details{
Similar in spirit to \code{\link[=consensus]{consensus()}} but for continuous covariates
instead of levels of a factor (and with a different usage).
}
\examples{
#load shells data
data("shells")

#compute shape expected at all the empiric values of size
mod <- lm(shells$shapes$coe ~ log(shells$sizes))
predshapes <- expected_shapes(mod, xvalue = log(shells$sizes))
pile_shapes(predshapes, mshape = FALSE)

#compute shape expected at the maximum and maximum sizes
mod <- lm(shells$shapes$coe ~ log(shells$sizes))
bigshape <- expected_shapes(mod, xvalue = max(log(shells$sizes)))
smallshape <- expected_shapes(mod, xvalue = min(log(shells$sizes)))

plot(inv_efourier(bigshape, 300), type = "l", lwd = 2)
plot(inv_efourier(smallshape, 300), type = "l", lwd = 2)
}
