% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hulls_by_group_3D}
\alias{hulls_by_group_3D}
\title{Plot 3D convex hulls for a series of groups}
\usage{
hulls_by_group_3D(xyz, fac, col = 1:nlevels(fac), ...)
}
\arguments{
\item{xyz}{Coordinates for the scatterplot}

\item{fac}{A factor grouping data points.}

\item{col}{A vector (either character or numeric) indicating the colors used
for each group.}

\item{...}{Further arguments passed to \code{\link[=rgl.triangles]{rgl.triangles()}} (e.g.
\code{specular}, \code{alpha}).}
}
\value{

}
\description{
Plot convex hulls for different groups in 3D scatterplots
created using \code{rgl}.
}
