% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{inv_efourier}
\alias{inv_efourier}
\title{Sample shapes along a morphometric axis}
\usage{
inv_efourier(coe, nb.p = 120)
}
\arguments{
\item{coe}{A vector with Fourier coefficients.}

\item{nb.p}{Numeric, specifying the number of coordinates for sampling the
outlines.}

\item{k}{}
}
\value{



}
\description{
A wrapper for \code{\link[=efourier_i]{efourier_i()}} from \code{Momocs} to transform a
set of Fourier coefficients into (x,y) coordinates. Used internally.
}
\examples{
Iverse Fourier transform

}
