% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{morphogrid}
\alias{morphogrid}
\title{Generate background shape models}
\usage{
morphogrid(
  ordination,
  axes,
  datype,
  template = NULL,
  x = NULL,
  y = NULL,
  p = NULL,
  k = NULL,
  nh = 5,
  nv = 4,
  mag = 1,
  asp = NA,
  xlim = NULL,
  ylim = NULL,
  rot.models = 0,
  size.models = 1,
  asp.models = 1
)
}
\arguments{
\item{ordination}{An ordination (i.e. a \code{"prcomp"}, \code{"bg_prcomp"},
\code{"phy_prcomp"} or \code{"pls"} object).}

\item{axes}{Numeric of length 1 or 2, indicating the morphometric axes to be
plotted. If values for either \code{x} or \code{y} are provided, only the
first value of this argument is considered.}

\item{datype}{type of shape data (either "fcoef" or "landm").}

\item{template}{A 2-column matrix containing 1) the actual
landmarks/semilandmarks being analized, followed by 2) the (x,y) cartesian
coordinates defining a curve or set of curves that will be warped using the
deformation interpolated from the changes between landmarks/semilandmarks
(the actual positions of which must be marked with a row of NA, see the
tails dataset).}

\item{x}{Optional vector with a non-morphometric variable to be plotted in
the x axis.}

\item{y}{Optional vector with a non-morphometric variable to be plotted in
the y axis.}

\item{p}{Numeric, indicating the number of landmarks/semilandmarks used (for
landmark data only).}

\item{k}{Numeric, indicating the number of cartesian dimensions of
landmarks/semilandmarks (for landmark data only).}

\item{nh}{Numeric; the number of shape models along the x axis.}

\item{nv}{Numeric; the number of shape models along the y axis.}

\item{mag}{Numeric; magnifying factor for shape models.}

\item{xlim, ylim, asp}{Standard arguments passed to the generic plot function}

\item{rot.models}{Numeric; angle (in degrees) to rotate shape models.}

\item{size.models}{Numeric; size factor for shape models.}

\item{asp.models}{Numeric; the y/x aspect ratio of shape models.}
}
\value{
A list of length 2 containing:
\itemize{
\item \code{$models_mat:} {a 2-column matrix with the (x,y) coordinates of
all the shape models in the background.}
\item \code{$models_arr:} {same as \code{models_mat} but in 3-margins array
format.}
}
}
\description{
Calculate and arrange background shape models for morphospaces.
For internal use.
}
\examples{
#perform pca on tails shapes
data("tails")
pca <- prcomp(geomorph::two.d.array(tails$shapes))

#generate grid of shape models sampling the range of variation
#at 4 locations (the 4 corners of the scatterplot)
shapes_grid <- morphogrid(ordination = pca, axes = c(1,2), datype = "landm",
                          k = ncol(tails$shapes), p = nrow(tails$shapes),
                          nh = 2, nv = 2)

#plot grid from $models_mat and project each shape in models_arr
plot(shapes_grid$models_mat)
points(shapes_grid$models_arr[,,1], pch=16, col = 1)
points(shapes_grid$models_arr[,,2], pch=16, col = 2)
points(shapes_grid$models_arr[,,3], pch=16, col = 3)
points(shapes_grid$models_arr[,,4], pch=16, col = 4)
}
