% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimality.R
\name{pareto_front}
\alias{pareto_front}
\title{Find Pareto front}
\usage{
pareto_front(performance, opt.increases, reverse = FALSE)
}
\arguments{
\item{performance}{No. observations-by-No. performance variables matrix of
performance data.}

\item{opt.increases}{Logical vector of length no. performance variables
indicating optimality direction of each performance variable.}

\item{reverse}{Logical indicating whether to reverse the order of output.}
}
\value{
Indices of Pareto optimal observations.
}
\description{
Find Pareto optimal observations in a set of performance data
given the optimality directions of each performance variable.
}
\details{
This function finds the Pareto Optimal Subset, or the Pareto front,
of a set of performance data.
}
\references{
Deakin, W. J., Anderson, P. S., den Boer, W., Smith, T. J., Hill, J. J.,
Rücklin, M., Donoghue, P. C. J.  & Rayfield, E. J. (2022). \emph{Increasing
morphological disparity and decreasing optimality for jaw speed and
strength during the radiation of jawed vertebrates}. Science Advances,
8(11), eabl3644.
}
\author{
William J. Deakin
}
