% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimality.R
\name{pareto_rank}
\alias{pareto_rank}
\title{Goldberg Pareto rank}
\usage{
pareto_rank(performance, opt.increases)
}
\arguments{
\item{performance}{No. observations-by-No. performance variables matrix of
performance data.}

\item{opt.increases}{Logical vector of length no. performance variables
indicating optimality direction of each performance variable.}
}
\value{
Integer Goldberg Pareto ranks of data.
}
\description{
Find the Goldberg Pareto rank of a set of performance data.
}
\details{
This function ranks performance data based on its optimality
direction. This is a Goldberg Pareto ranking, which iteratively finds the
Pareto front, assigns it an integer rank, then removes it from the dataset
and repeats.
}
\references{
Deakin, W. J., Anderson, P. S., den Boer, W., Smith, T. J., Hill, J. J.,
Rücklin, M., Donoghue, P. C. J.  & Rayfield, E. J. (2022). \emph{Increasing
morphological disparity and decreasing optimality for jaw speed and
strength during the radiation of jawed vertebrates}. Science Advances,
8(11), eabl3644.
}
\author{
William J. Deakin
}
