% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimality.R
\name{pareto_rank_ratio}
\alias{pareto_rank_ratio}
\title{Pareto Rank Ratio}
\usage{
pareto_rank_ratio(performance, opt.increases)
}
\arguments{
\item{performance}{No. observations-by-No. performance variables matrix of
performance data.}

\item{opt.increases}{Logical vector of length no. performance variables
indicating optimality direction of each performance variable.}
}
\value{
Numeric vector containing the Pareto Rank Ratio of the performance
data.
}
\description{
Find the Pareto rank ratio of a set of performance data.
}
\details{
This function uses the formula from Deakin et al. (2022) to
calculate the Pareto rank ratio of the performance data.
}
\references{
Deakin, W. J., Anderson, P. S., den Boer, W., Smith, T. J., Hill, J. J.,
Rücklin, M., Donoghue, P. C. J.  & Rayfield, E. J. (2022). \emph{Increasing
morphological disparity and decreasing optimality for jaw speed and
strength during the radiation of jawed vertebrates}. Science Advances,
8(11), eabl3644.
}
\author{
William J. Deakin
}
