% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{phyalign_comp}
\alias{phyalign_comp}
\title{Phylogenetically Aligned Component Analysis}
\usage{
phyalign_comp(x, tree, corr = FALSE)
}
\arguments{
\item{x}{A matrix with one or more variables as columns and observations as
rows. Row must be named and match tip labels from the phylogenetic
\code{tree}.}

\item{tree}{A \code{"phylo"} object containing a phylogenetic tree. Tip
labels should match the row names from \code{x}.}
}
\description{
Performs Phylogenetically Aligned Component Analysis (PACA).
Experimental.
}
