% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pile_shapes}
\alias{pile_shapes}
\title{Pile shapes}
\usage{
pile_shapes(shapes, links = NULL, mshape = TRUE, ...)
}
\arguments{
\item{shapes}{Shape data.}

\item{links}{An optional list with the indices of the coordinates defining
the wireframe (following the format used in \code{Morpho}).}

\item{mshape}{Logical; whether to plot the mean configuration.}

\item{...}{Additional arguments passed to \code{\link[=plot]{plot()}}.}
}
\value{

}
\description{
Superimpose all the shapes in the sample.
}
