% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_biv_landscape}
\alias{plot_biv_landscape}
\title{Plot bivariate landscape into morphospace}
\usage{
plot_biv_landscape(
  landscape,
  display,
  type,
  levels,
  lwd,
  lty,
  col,
  drawlabels,
  alpha
)
}
\arguments{
\item{landscape}{A list containing the results of \code{\link[akima:interp]{akima::interp()}}.}

\item{display}{Either \code{"contour"} or \code{"filled.contour"}.}

\item{type}{Type of landscape to be plotted. Options are \code{"theoretical"}
or \code{"empirical"}.}

\item{levels}{Levels to be used to create the contours.}

\item{lwd}{Numeric; width of the lines depicting contours.}

\item{lty}{Numeric; type of the lines depicting contours.}

\item{col}{Colors used to represent the landscape contours.}

\item{drawlabels}{Logical; should the labels indicating the value of each
surface contour be plotted?}

\item{alpha}{Transparency factor for filled contours.}
}
\description{
An \code{ad hoc} wrapper for \code{\link[graphics:contour]{graphics::contour()}} /
\code{\link[graphics:filled.contour]{graphics::.filled.contour()}} /  \code{\link[graphics:image]{graphics::image()}}. Used internally for
projecting landscapes into bivariate morphospaces.
}
