% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_biv_scatter}
\alias{plot_biv_scatter}
\title{Plot scatterpoints into bivariate morphospace}
\usage{
plot_biv_scatter(scores, col = 1, bg = 1, pch = 1, cex = 1, ...)
}
\arguments{
\item{scores}{a matrix of (x,y) coordinates to plot.}

\item{col}{Color of the scatterpoints.}

\item{bg}{Background color for the scatterpoints.}

\item{pch}{Symbol of the scatterpoints.}

\item{cex}{Numeric; size of the scatterpoints.}

\item{...}{Further arguments passed to \code{\link[graphics:points]{graphics::points()}}.}
}
\description{
An \code{ad hoc} wrapper for \code{\link[graphics:points]{graphics::points()}}. Used
internally for projecting scatterpoints into bivariate morphospaces.
}
