% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{plot_morphogrid2d}
\alias{plot_morphogrid2d}
\title{Plot background 2D shape models}
\usage{
plot_morphogrid2d(
  x = NULL,
  y = NULL,
  morphogrid,
  template = NULL,
  links = NULL,
  datype,
  ordtype,
  axes,
  adj_frame = c(1, 1),
  p,
  xlab = NULL,
  ylab = NULL,
  cex.ldm,
  col.ldm,
  col.models,
  lwd.models,
  bg.models,
  plot = TRUE
)
}
\arguments{
\item{x}{Optional vector with a non-morphometric variable to be plotted in
the x axis.}

\item{y}{Optional vector with a non-morphometric variable to be plotted in
the y axis.}

\item{morphogrid}{An object containing the output of \code{morphogrid}.}

\item{template}{A 2-column matrix containing landmarks/semilandmarks followed
by coordinates defining a curve or set of curves describing additional aspects
of morphology, which will be warped using TPS interpolation to produce the set
of background shell models(see \code{\link{build_template2d}}).}

\item{links}{A list with the indices of the coordinates defining the
wireframe (following the format used in \code{Morpho}).}

\item{datype}{Character; type of shape data used (\code{"landm"} or
\code{"fcoef"}).}

\item{ordtype}{Character; method used for multivariate ordination
(\code{"prcomp"}, \code{"bg_prcomp"}, \code{"phy_prcomp"}, \code{"pls_shapes"}
or \code{"phy_pls_shapes"}).}

\item{axes}{Numeric of length 2, indicating the axes to be plotted.}

\item{adj_frame}{Numeric of length 2, providing \emph{a posteriori} scaling
factors for the width and height of the frame, respectively.}

\item{p}{Numeric, indicating the number of landmarks/semilandmarks used (for
landmark data only).}

\item{xlab, ylab}{Standard arguments passed to the generic plot function.}

\item{cex.ldm}{Numeric; size of landmarks/semilandmarks in the background
models.}

\item{col.ldm}{The color of landmarks/semilandmarks in the background models.}

\item{col.models}{The color for wireframes/outlines.}

\item{lwd.models}{Numeric; the width of the lines in wireframes/outlines.}

\item{bg.models}{Background color for outlines.}

\item{plot}{Logical; whether to plot morphospace.}
}
\description{
Plot the output from \code{\link[=morphogrid]{morphogrid()}}, for 2-dimensional morphometric
data. Used internally.
}
\examples{
#load data and packages
library(geomorph)
data("tails")

#perform pca on tails shapes
pca <- prcomp(two.d.array(tails$shapes))

#generate grid of shape models sampling the range of variation
#at 4 locations (the 4 corners of the scatterplot)
shapes_grid <- morphogrid(ordination = pca, axes = c(1,2), datype = "landm",
                          k = ncol(tails$shapes), p = nrow(tails$shapes),
                          nh = 2, nv = 2)

#plot grid
plot_morphogrid2d(morphogrid = shapes_grid, links = tails$links,
                  datype = "landm", ordtype = "prcomp",
                  axes = c(1,2), p = 9, col.ldm = 1, cex.ldm = 1,
                  col.models = 1, lwd.models = 1)

#project each shape in models_arr
points(shapes_grid$models_arr[,,1], pch=16, col = 1)
points(shapes_grid$models_arr[,,2], pch=16, col = 2)
points(shapes_grid$models_arr[,,3], pch=16, col = 3)
points(shapes_grid$models_arr[,,4], pch=16, col = 4)
}
\seealso{
\code{\link{morphogrid}}, \code{\link{adjust_models2d}}
}
