% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{plot_morphogrid3d}
\alias{plot_morphogrid3d}
\title{Plot background 3D shape models}
\usage{
plot_morphogrid3d(
  x = NULL,
  y = NULL,
  morphogrid,
  refshape,
  template = NULL,
  links = NULL,
  ordtype,
  axes,
  p,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  cex.ldm,
  col.ldm,
  col.models,
  lwd.models,
  size.models,
  alpha.models = 1,
  bg.models,
  asp.models,
  plot = TRUE
)
}
\arguments{
\item{x}{Optional vector with a non-morphometric variable to be plotted in
the x axis.}

\item{y}{Optional vector with a non-morphometric variable to be plotted in
the y axis.}

\item{morphogrid}{An object containing the output of \code{morphogrid}.}

\item{refshape}{reference shape (i.e., the mean landmark configuration)
corresponding to the mesh provided in \code{template}.}

\item{template}{An optional \code{"mesh3d"} object, which will be warped using
TPS interpolation to produce the set of background shell models.}

\item{links}{A list with the indices of the coordinates defining the
wireframe (following the format used in \code{Morpho}).}

\item{ordtype}{Character; method used for multivariate ordination
(\code{"prcomp"}, \code{"bg_prcomp"}, \code{"phy_prcomp"}, \code{"pls_shapes"}
or \code{"phy_pls_shapes"}).}

\item{axes}{Numeric of length 2, indicating the axes to be plotted.}

\item{p}{Numeric, indicating the number of landmarks/semilandmarks used (for
landmark data only).}

\item{xlim, ylim, xlab, ylab}{Standard arguments passed to the generic plot
function.}

\item{cex.ldm}{Numeric; size of landmarks/semilandmarks in the background
models.}

\item{col.ldm}{The color of landmarks/semilandmarks in the background models.}

\item{col.models}{The color for wireframes/outlines.}

\item{lwd.models}{Numeric; the width of the lines in wireframes/outlines.}

\item{size.models}{Numeric; size factor for shape models.}

\item{alpha.models}{Numeric; transparency factor for background models.}

\item{bg.models}{Background color for outlines.}

\item{asp.models}{Numeric; the y/x aspect ratio of shape models.}

\item{plot}{Logical; whether to plot morphospace.}
}
\description{
Plot the output from \code{\link[=morphogrid]{morphogrid()}}, for 3-dimensional morphometric
(landmark) data. Used internally.
}
\details{
This function allows the user to choose the orientation of the 3D
models by interactively rotating a shape model. Do not close the \code{rgl}
window, nor minimize it actively (just bring back Rstudio to the front and
let the device get minimized pasively). The process of generating the morphospace
is rather slow, specially if a mesh is provided for \code{template} and/or if
\code{alpha.models} value is lower than \code{1}.
}
