% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{plot_mspace}
\alias{plot_mspace}
\title{Plot morphospaces and combine them with other variables}
\usage{
plot_mspace(
  mspace,
  axes = NULL,
  links = NULL,
  template = NULL,
  x = NULL,
  y = NULL,
  nh,
  nv,
  mag,
  invax = NULL,
  adj_frame = c(1, 1),
  points = TRUE,
  models = TRUE,
  mshapes = TRUE,
  groups = TRUE,
  phylo = TRUE,
  shapeax = TRUE,
  landsc = TRUE,
  legend = FALSE,
  cex.legend = 1,
  asp = NA,
  xlab,
  ylab,
  xlim = NULL,
  ylim = NULL,
  size.models,
  asp.models,
  rot.models,
  col.models,
  bg.models,
  lwd.models,
  alpha.models,
  cex.ldm,
  col.ldm,
  pch.points = 1,
  col.points = 1,
  bg.points = 1,
  cex.points = 1,
  density.points = TRUE,
  col.groups = 1,
  bg.groups = 1,
  pch.groups = 16,
  cex.groups = 1,
  ellipse.groups = mspace$plotinfo$ellipse.groups,
  conflev.groups = 0.95,
  lwd.groups = 1,
  lty.groups = 1,
  alpha.groups = 0,
  density.groups = TRUE,
  lwd.phylo = 1,
  lty.phylo = 1,
  col.phylo = 1,
  lwd.axis = 1,
  lty.axis = 1,
  col.axis = 1,
  palette.landsc = heat.colors,
  ncols.landsc = 50,
  nlevels.landsc = 50,
  drawlabels.landsc = FALSE,
  lty.landsc = 1,
  lwd.landsc = 1
)
}
\arguments{
\item{mspace}{An \code{"mspace"} object created using \code{\link[=mspace]{mspace()}}.}

\item{axes}{Numeric of length 1 or 2, indicating the axes to be plotted.}

\item{links}{A list with the indices of the coordinates defining the
wireframe (following the format used in \code{Morpho}).}

\item{template}{Either a 2-column matrix containing landmarks/semilandmarks
followed by coordinates defining a curve or set of curves describing additional
aspects of morphology (for 2D shape data) or a \code{"mesh3d"} object containing
geometry of the structure the landmarks were placed on (for 3D shape data),
corresponding to the mean shape of the sample. These will be warped using TPS
interpolation to produce the set of background shell models (see
\code{\link{build_template2d}}).}

\item{x}{Optional vector with a non-morphometric variable to be plotted in
the x axis. Alternatively, a \code{"phylo"} object can be provided.}

\item{y}{Optional vector with a non-morphometric variable to be plotted in
the y axis. Alternatively, a \code{"phylo"} object can be provided.}

\item{nh}{Numeric; the number of shape models along the x axis.}

\item{nv}{Numeric; the number of shape models along the y axis.}

\item{mag}{Numeric; magnifying factor for shape models.}

\item{invax}{Optional numeric indicating which of the axes provided in
\code{axes} needs to be inverted (optionsare \code{1}, \code{2} or
\code{c(1,2)}).}

\item{adj_frame}{Numeric of length 2, providing \emph{a posteriori} scaling
factors for the width and height of the frame, respectively.}

\item{points}{Logical; whether to plot the scatter points corresponding to
the sampled shapes stored in \code{mspace$x}.}

\item{models}{Logical; whether to plot background shape models.}

\item{mshapes}{Logical; whether to plot the scatter points corresponding to
groups' mean shapes stored in \code{mspace$gr_centroids}.}

\item{groups}{Logical; whether to plot the convex hulls/confidence ellipses
enclosing the groups stored in \code{mspace$gr_class}.}

\item{phylo}{Logical; whether to plot phylogenetic relationships stored
in \code{mspace$phylo}.}

\item{shapeax}{Logical; whether to plot morphometric axes stored in
\code{mspace$shape_axis}.}

\item{landsc}{Logical; whether to plot landscape surface stored in
\code{mspace$landsc}.}

\item{legend}{Logical; whether to show legend for groups (\code{mspace$gr_class}).}

\item{cex.legend}{Numeric; size of legend labels/symbols.}

\item{xlim, ylim, xlab, ylab, asp}{Standard arguments passed to the generic \code{\link[graphics:plot.default]{graphics::plot()}}
function.}

\item{size.models}{Numeric; size factor for shape models.}

\item{asp.models}{Numeric; y/x aspect ratio of shape models.}

\item{rot.models}{Numeric; angle (in degrees) to rotate shape models.}

\item{col.models}{Color for wireframes/outlines.}

\item{bg.models}{Background color for outlines/meshes.}

\item{lwd.models}{Numeric; width of the lines in wireframes/outlines.}

\item{alpha.models}{Numeric; transparency factor for background models (3D only).}

\item{cex.ldm}{Numeric; size of landmarks/semilandmarks in the background
models.}

\item{col.ldm}{Color of landmarks/semilandmarks in the background models.}

\item{pch.points}{Numeric; symbol of the scatter points corresponding to
sampled shapes.}

\item{col.points}{Color of the scatter points corresponding to sampled
shapes.}

\item{bg.points}{Background color of the scatter points corresponding
to sampled shapes.}

\item{cex.points}{Numeric; size of the scatter points corresponding to
sampled shapes.}

\item{density.points}{Logical; whether to add density distribution for points
(univariate ordinations only).}

\item{col.groups}{Color of the hulls/ellipses and/or scatter points corresponding
to groups mean shapes.}

\item{bg.groups}{Background color of the scatter points corresponding
to groups mean shapes.}

\item{pch.groups}{Numeric; symbol of the scatter points corresponding to
groups mean shapes.}

\item{cex.groups}{Numeric; size of the scatter points corresponding to
groups mean shapes.}

\item{ellipse.groups}{Logical; whether to plot confidence ellipses (if \code{FALSE},
convex hulls will be used instead).}

\item{conflev.groups}{Numeric; confidence level used for confidence ellipse(s).}

\item{lwd.groups}{Numeric; width of the lines in groups' ellipses/hulls.}

\item{lty.groups}{Numeric; type of the lines in groups' ellipses/hulls.}

\item{alpha.groups}{Numeric; transparency factor for groups' ellipses/hulls/density
distributions.}

\item{density.groups}{Logical; whether to add density distribution for groups
(univariate ordinations only).}

\item{lwd.phylo}{Numeric; width of the lines depicting phylogenetic
branches.}

\item{lty.phylo}{Numeric; type of the lines depicting phylogenetic branches.}

\item{col.phylo}{Numeric; color of the lines depicting phylogenetic branches.}

\item{lwd.axis}{Numeric; width of the lines depicting morphometric axis.}

\item{lty.axis}{Numeric; type of the lines depicting  morphometric axis.}

\item{col.axis}{Numeric; color of the lines depicting morphometric axis.}

\item{palette.landsc}{Color palette to use for landscape representation.}

\item{ncols.landsc}{Number of colors to use for landscape representation.}

\item{nlevels.landsc}{Number of levels (i.e. contours) to use in landscape representation.}

\item{drawlabels.landsc}{Logical; should the labels indicating the value of each surface contour be plotted?}

\item{lty.landsc}{Numeric; type of the lines depicting surface contour.}

\item{lwd.landsc}{Numeric; width of the lines depicting surface contour.}
}
\description{
Flexible representation of morphospaces, including their combination
with other variables or a phylogeny.
}
\details{
This function allows to regenerate morphospaces contained in \code{mspace}
objects already in existence, either preserving the graphical attributes used
during the pipeline or modifying one or more of them (so there is no need to
execute the pipeline every time morphospaces need to be visualized and/or changed).

Also, \code{plot_mspace} expands the range of graphical options available
beyond 'pure' morphospaces. If a numeric non-shape variable (assumed to have
been measured for the same specimens in \code{mspace$x}) is feeded to one of
\code{x} or \code{y}, a 'hybrid' morphospace is produced (i.e. the bivariate
plot will be constructed from the combination of \code{x} or \code{y} and a
morphometric axis; shape models in the background will represent variation
only for the latter). If instead a \code{"phylo"} object (assumed to describe
the phylogenetic relationships among tips scores stored in
\code{mspace$phylo_scores}) is feeded to one of \code{x} or \code{y}, a vertical
or horizontal phenogram will be deployed (the x/y axis range will correspond to
branch lengths, so caution should be exercised when interpreting the output).
}
\examples{
#load and extract relevant data, packages and information
library(magrittr)
data("tails")
shapes <- tails$shapes
species <- tails$data$species
sizes <- tails$sizes
sp_shapes <- expected_shapes(shapes, species)
sp_sizes <- cbind(tapply(sizes, species, mean))
tree <- tails$tree
links <- tails$links

#generate basic morphospace, add sampled shapes, species mean shapes, species
#classification, and phylogenetic structure
msp <- mspace(shapes, axes = c(1,2), plot = FALSE) \%>\%
 proj_shapes(shapes = shapes) \%>\%
 proj_consensus(shapes = sp_shapes) \%>\%
 proj_groups(groups = species) \%>\%
 proj_phylogeny(tree = tree)

##Plotting 'pure' morphospaces:

#plot mspace object as it is
plot_mspace(msp, axes = c(1,2),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#add colors for points, by species
plot_mspace(msp, axes = c(1,2), col.points = species, col.groups = 1:nlevels(species),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#add links
plot_mspace(msp, axes = c(1,2), links = links,
            col.points = species, col.groups = 1:nlevels(species),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#change number and sizes of shape models in the background
plot_mspace(msp, axes = c(1,2), nh = 2, nv = 2, links = links, size.models = 0.5,
            col.points = species, col.groups = 1:nlevels(species),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#magnify deformation and highlight landmarks
plot_mspace(msp, axes = c(1,2), mag = 1.5, nh = 2, nv = 2, links = links, size.models = 0.5,
            col.points = species, col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#change axes 1,2 for 1,3
plot_mspace(msp, axes = c(1,3), mag = 1.5, nh = 2, nv = 2, links = links, size.models = 0.5,
            col.points = species, col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#represent ranges and centroids of groups
plot_mspace(msp, axes = c(1,3), mag = 1.5, nh = 2, nv = 2, links = links, size.models = 0.5,
            col.points = species, col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = TRUE, mshapes = TRUE, phylo = FALSE)

#add phylogeny
plot_mspace(msp, axes = c(1,3), mag = 1.5, nh = 2, nv = 2, links = links, size.models = 0.5,
            col.points = species, col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = TRUE, mshapes = TRUE, phylo = TRUE)



##Plotting 'hybrid' morphospaces:

#plot size against first PC
plot_mspace(msp, x = sizes,  axes = 1, links = links, col.points = species,
            col.groups = 1:nlevels(species), pch.points = 16,
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE, xlab = "Centroid size")



##Plotting phenograms:

#plot horizontal phenogram against PC1
plot_mspace(msp, x = tree,  axes = 1, links = links, col.points = species,
            col.groups = 1:nlevels(species), pch.points = 16,
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE, xlab = "Branch lengths")

#plot horizontal phenogram against PC1
plot_mspace(msp, y = tree,  axes = 2, links = links, col.points = species,
            col.groups = 1:nlevels(species), pch.points = 16,
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE, ylab = "Branch lengths")
}
