% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_univ_landscape}
\alias{plot_univ_landscape}
\title{Plot univariate landscape into morphospace}
\usage{
plot_univ_landscape(landscape, drawlabels, col, lwd)
}
\arguments{
\item{landscape}{A list containing the results of \code{\link[akima:interp]{akima::interp()}}.}

\item{drawlabels}{Logical; should the labels indicating the value of each
surface contour be plotted?}

\item{col}{Colors used to represent the landscape curve.}

\item{lwd}{Numeric; width of the lines depicting the landscape curve.}
}
\description{
Plot a curve representing a "univariate" landscape. Used
internally for projecting landscapes into univariate morphospaces.
}
