% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_univ_scatter}
\alias{plot_univ_scatter}
\title{Plot scatterpoints into univariate morphospace}
\usage{
plot_univ_scatter(scores, density, col = 1, bg = 1, pch = 1, cex = 1, ...)
}
\arguments{
\item{scores}{a matrix of (x,y) coordinates to plot.}

\item{density}{Logical; whether to plot density curve.}

\item{col}{Numeric; color of the scatterpoints.}

\item{bg}{Numeric; background color for the scatterpoints.}

\item{pch}{Numeric; symbol of the scatterpoints.}

\item{cex}{Numeric; size of the scatterpoints.}

\item{...}{Further arguments passed to \code{\link[graphics:points]{graphics::points()}}.}
}
\description{
An \code{ad hoc} wrapper for \code{\link[stats:density]{stats::density()}} +
\code{\link[graphics:polygon]{graphics::polygon()}} / \code{\link[graphics:points]{graphics::points()}}. Used internally for
projecting scatterpoints into univariate morphospaces.
}
