% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{pls2b}
\alias{pls2b}
\title{Two-blocks Partial Least Squares}
\usage{
pls2b(x, y, LOOCV = FALSE, recompute = FALSE)
}
\arguments{
\item{x}{A matrix with one or more variables as columns and observations as
rows, representing the first block.}

\item{y}{A matrix with one or more variables as columns and observations as
rows, representing the second block.}

\item{LOOCV}{Logical; whether to apply leave-one-out cross-validation.}

\item{recompute}{Logical; whether to re-compute rotation matrix using the
scores resulting from LOOCV.}
}
\value{
A \code{"pls2b"} object, containing:
\itemize{
\item \code{$values:} {vector of singular values accounting for the
covariation among blocks explained by each par of axes.}
\item \code{$xrotation:} {matrix of vector coefficients for the first
block.}
\item \code{$yrotation:} {matrix of vector coefficients for the second
block.}
\item \code{$xcenter:} {the mean values of the original variables from
the first block.}
\item \code{$ycenter:} {the mean values of the original variables from
the second block.}
\item \code{$xtotvar:} {the sum of the variances from the original
variables from the second block.}
\item \code{$ytotvar:} {the sum of the variances from the original
variables from the second block.}
}
}
\description{
Performs 2B Partial Least Squares allowing for leave-one-out
cross-validation, which is useful one the number of variables exceeds the
number of observations (i.e. alleviates spurious covariation between
variables).
}
\details{
Starting with two blocks of variables measured for the same
cases, two-blocks PLS finds the linear combination of variables on each
block maximizing covariation with the variables on the other. In the
context of \code{morphospace}, one or both of these blocks will usually
be a series of shapes arranged as a 2-margins matrix.

It has been reported that PLS (as an algebraic equivalent of bgPCA)
produces spurious covariation between blocks when the number of variables
exceeds the number of observations (which is a common situation in
geometric morphometrics analyses). This problem can be alleviated by
carrying out a leave-one-out cross-validation (LOOCV; i.e. each
observation is excluded from the calculation of PLS axes before its
projection in the resulting ordination as a way to calculate its score).
}
\examples{
#load data
data("tails")

#extract shapes and log sizes
shapes <- tails$shapes
sizes <- log(tails$sizes)

#perform PLS between shape and size
pls <- pls2b(y = geomorph::two.d.array(shapes), x = sizes)

#look at the results
names(pls) #the contents of the resulting object
plot(pls$xscores, pls$yscores) #ordination

#pls_shapes achieves identical results, but it is formatted to be used more
#easily when analyzing shape variables and its output is compatible with other
#functions from morphospace
pls2 <- pls_shapes(shapes = shapes, X = sizes)

names(pls2) #the contents of the resulting object
exp_var(pls2) #variance explained by each axis
plot(pls2$x2, pls2$x) #ordination
}
\references{
Rohlf, F. J., & Corti, M. (2000). \emph{Use of two-block partial
least-squares to study covariation in shape}. Systematic Biology, 49(4),
740-753.

Zelditch, M. L., Swiderski, D. L., & Sheets, H. D. (2012).
\emph{Geometric morphometrics for biologists: A primer}. 2nd ed. Academic
Press.

Bookstein, F. L. (2019). \emph{Pathologies of between-groups principal
components analysis in geometric morphometrics}. Evolutionary Biology, 46(4),
271-302.
}
\seealso{
\code{\link{pls_shapes}}, \code{\link{phy_pls2b}},
\code{\link{exp_var}}
}
