% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{points.mspace}
\alias{points.mspace}
\title{Plot points for mspace objects}
\usage{
\method{points}{mspace}(mspace, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot points for mspace objects
}
