% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{print.mspace}
\alias{print.mspace}
\title{Print \code{"mspace"} objects}
\usage{
\method{print}{mspace}(mspace)
}
\arguments{
\item{mspace}{An \code{"mspace"} object.}
}
\description{
Print contents of \code{"mspace"} objects in a friendly way :)
}
\details{
Will return information about how the ordination was obtained
(method and data) as well as the elements projected into it.
}
