% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_axes}
\alias{proj_axes}
\title{Project morphometric axis/axes into morphospace}
\usage{
proj_axes(neword, mspace, ax = 1, p = NULL, k = NULL, pipe = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Project morphometric axis/axes into morphospace
}
