% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_axis}
\alias{proj_axis}
\title{Project morphometric axis into morphospace}
\usage{
proj_axis(obj, mspace, axis = 1, mag = 1, pipe = TRUE, ...)
}
\arguments{
\item{obj}{An object containing either a multivariate ordination of class
\code{"prcomp", "bg_prcomp", "phy_prcomp"} or \code{"pls_shape"} or a
\code{"mlm"} object fitted using \code{\link[=lm]{lm()}}.}

\item{mspace}{An \code{"mspace"} object created using \code{\link[=mspace]{mspace()}}.}

\item{axis}{An optional vector indicating the axis from \code{obj} to be
projected.}

\item{mag}{Numeric; magnifying factor for representing shape transformation.}

\item{pipe}{Logical; is the function being included in a pipe?}

\item{...}{Further arguments passed to \code{\link[=lines]{lines()}}.}
}
\value{

}
\description{
Project one or more morphometric axes (i.e., linear combinations
of shape variables) into an existing morphospace.
}
