% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_consensus}
\alias{proj_consensus}
\title{Project consensus shape(s) into morphospace}
\usage{
proj_consensus(consensus, mspace, add = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Project consensus shape(s) into morphospace
}
