% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_builders.R
\name{proj_eigen}
\alias{proj_eigen}
\title{Project cases into existing eigenanalysis-based ordination}
\usage{
proj_eigen(x, vectors)
}
\arguments{
\item{vectors}{}
}
\value{

}
\description{
Project cases into existing eigenanalysis-based ordination
}
