% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_pfront}
\alias{proj_pfront}
\title{Project Pareto front optimizing multiple functions}
\usage{
proj_pfront(
  mspace,
  shapes = NULL,
  X = NULL,
  opt.increases = NULL,
  lwd = 2,
  pipe = TRUE,
  ...
)
}
\arguments{
\item{mspace}{An \code{"mspace"} object.}

\item{shapes}{Optional shape data. If provided, the associated variables
must be specified using \code{X}.}

\item{X}{An optional matrix with at least 2 columns, containing the values
assigned to each shape (if provided) in the functional variables the Pareto
is calculated from (vector length / number of rows and their order must
match those from the shapes provided in \code{shapes}).}

\item{opt.increases}{Logical vector of length \code{o} (where \code{o} is the
number of \emph{ad hoc} functions or variables provided in \code{FUN} or
\code{X}, respectively) indicating optimality direction of each performance
variable. Ignored unless projection of an optimality trade-off landscape is
attempted.}

\item{lwd}{Numerical, specifying the width of the line used to mark
the Pareto front.}

\item{pipe}{Logical; is the function being included in a pipe?}

\item{...}{Further arguments passed to \code{\link[graphics:lines]{graphics::lines()}}.}
}
\value{
If a plot device with a morphospace is open, the Pareto front is
projected into it. If \code{pipe = FALSE}, a matrix containing the shapes
forming the Pareto front in its "two dimensional" format is returned
invisibly. If \code{pipe = TRUE} the supplied \code{"mspace"} object will
be modified by appending \code{$pfront} and \code{$optimality} slots to
\code{$projected}, as well as by adding some graphical parameters (stored
into the \code{$plotinfo} slot), and returned invisibly.
}
\description{
Project the Pareto front, representing the subset of shapes that
achieve the optimal balance between two antagonistic functional variables,
into morphospace.
}
\details{
This function projects the Pareto front, formed by the subset of
shapes that achieve a better balance between two or more functional metrics
engaged in a trade-off that cannot be maximised independently, following
Deakin et al. (2022)
}
\examples{
#load data and packages
library(geomorph)

data("tails")
shapes <- tails$shapes
links <- tails$links


#Pareto landscapes are created when providing two or more separate surfaces,
#either using a list fed to FUN containing multuple functions, or a matrix with
#multiple columns.
msp <- mspace(shapes, links = links, nh = 8, nv = 8, plot = FALSE)
shapemodels2d <- two.d.array(extract_shapes(msp, keep.template = FALSE)$shapes)
LDs <- apply(X = shapemodels2d, FUN = morphospace:::computeLD, MARGIN = 1)

#create fake functional variable with the opposite pattern to LD
antiLDs <- jitter(rev(LDs), factor = 10)

#visualize lift/drag landscape (aerodynamic efficiency)
mspace(shapes, links = links, nh = 8, nv = 8, size.model = 1.5,
       cex.ldm = 0) \%>\%
  proj_landscape(shapes = shapemodels2d, X = LDs,
                 nlevels = 20, expand = 1.2, lwd = 3,
                 palette = terrain.colors, display = "contour")

#visualize antiLD surface
mspace(shapes, links = links, nh = 8, nv = 8, size.model = 1.5,
       cex.ldm = 0) \%>\%
  proj_landscape(shapes = shapemodels2d, X = antiLDs,
                 nlevels = 20, expand = 1.2, lwd = 3,
                 palette = terrain.colors, display = "contour")

#visualize combined Pareto landscape (trade-off)
mspace(shapes, links = links, nh = 8, nv = 8, size.model = 1.5,
       cex.ldm = 0) \%>\%
  proj_landscape(shapes = shapemodels2d, X = cbind(LDs, antiLDs),
                 opt.increases = c(TRUE, TRUE),
                 nlevels = 50, expand = 1.2, lwd = 3,
                 palette = terrain.colors, display = "contour")


#there are two ways in which proj_pfront can be used:

#1. downstream to a proj_landscape instance in which a Pareto
#surface was computed
mspace(shapes, links = links, nh = 8, nv = 8, size.model = 1.5,
       cex.ldm = 0) \%>\%
  proj_landscape(shapes = shapemodels2d, X = cbind(LDs, antiLDs),
                 opt.increases = c(TRUE, TRUE),
                 nlevels = 50, expand = 1.2, lwd = 3,
                 palette = terrain.colors, display = "contour") \%>\%
  proj_shapes(shapemodels2d) \%>\%
  proj_pfront(col = "red")

#2. without previous computation of a Pareto surface. In this case,
#the shapes, functional values and directions of optimality must be
#provided
mspace(shapes, links = links, nh = 8, nv = 8, size.model = 1.5,
       cex.ldm = 0) \%>\%
  proj_pfront(shapes = shapemodels2d, X = cbind(LDs, antiLDs),
              opt.increases = c(TRUE, TRUE), bg = "red", type = "b",
              pch = 21)
}
\references{
Deakin, W. J., Anderson, P. S., den Boer, W., Smith, T. J., Hill, J. J.,
Rücklin, M., Donoghue, P. C. J.  & Rayfield, E. J. (2022). \emph{Increasing
morphological disparity and decreasing optimality for jaw speed and
strength during the radiation of jawed vertebrates}. Science Advances,
8(11), eabl3644.
}
\seealso{
\code{\link{pareto_front}}, \code{\link{proj_landscape}}
}
