% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_phylogeny}
\alias{proj_phylogeny}
\title{Project phylogenetic structure into morphospace}
\usage{
proj_phylogeny(
  mspace,
  shapes = NULL,
  tree,
  evmodel = "BM",
  labels.tips = FALSE,
  labels.nodes = FALSE,
  pch.nodes = 16,
  col.nodes = "gray",
  bg.nodes = 1,
  cex.nodes = 0.8,
  pch.tips = 16,
  bg.tips = 1,
  col.tips = 1,
  cex.tips = 1,
  pipe = TRUE,
  ...
)
}
\arguments{
\item{mspace}{An \code{"mspace"} object.}

\item{shapes}{Shape data, with 3rd margin names matching tip labels
from \code{tree}.}

\item{tree}{A \code{"phylo"} object containing a phylogenetic tree.}

\item{evmodel}{Character, specifying an evolutionary model to perform
ancestral character reconstruction; options are "BM" (Brownian motion;
implemented through \code{\link[phytools]{fastAnc}}), "EB" (Early burst),
"OU" (Ornstein-Uhlenbeck) and "lambda" (Pagel's lambda transformation)
(implemented using \code{\link[mvMORPH]{mvgls}}).}

\item{labels.tips}{Either logical, indicating whether to include tip labels,
or a character string with the exact names of the tips whose labels
should be included.}

\item{labels.nodes}{Either logical, indicating whether to include node
labels, or a character string with the exact names of the nodes whose
labels should be included (with the form of "node_n", e.g., "node_14"
corresponds to the root of a tree with 13 tips).}

\item{pch.nodes}{Symbol of the scatter points corresponding to the
nodes of the phylogeny.}

\item{col.nodes}{Color of the hulls/ellipses and/or scatter points
corresponding to the nodes of the phylogeny.}

\item{bg.nodes}{Background color of the scatter points corresponding to the
nodes of the phylogeny.}

\item{cex.nodes}{Numeric; size of the scatter points corresponding to the
nodes of the phylogeny.}

\item{pch.tips}{Symbol of the scatter points corresponding to the tips of the
phylogeny.}

\item{bg.tips}{Background color of the scatter points corresponding to the
tips of the phylogeny.}

\item{col.tips}{Color of the hulls/ellipses and/or scatter points
corresponding to the tips of the phylogeny.}

\item{cex.tips}{Numeric; size of the scatter points corresponding to the
tips of the phylogeny.}

\item{pipe}{Logical; is the function being included in a pipe?}

\item{...}{Further arguments passed to \code{\link[graphics]{lines}}
(commonly, \code{pch}, \code{col}/\code{bg} and \code{cex}.}
}
\value{
If a plot device with a morphospace is open, shapes representing the
tips and nodes of the phylogenetic tree, as well as the lines connecting
them, are projected into morphospace. If \code{pipe = FALSE} scores for
nodes and tips of the phylogeny are returned invisibly.
If \code{pipe = TRUE} the supplied \code{"mspace"} object will be modified
by appending \code{$phylo_scores}, \code{$phylo_evmodel}, \code{$phylo}
slots to \code{$projected}, as well as by adding some graphical parameters
(stored into the \code{$plotinfo} slot), and returned invisibly.
}
\description{
Project phylogenetic relationships among a set of shapes
(representing the tips of a phylogenetic tree) into an existing bivariate
morphospace.
}
\details{
The purpose of this function is twofold. First, it is meant to
transform a morphospace into a phylomorphospace by projecting node shapes
and phylogenetic relationships. To this end, a set of named shapes must be
provided; \code{dim(shapes)[3]} must match \code{tree$tip.labels}. Second,
this function can be used to retrieve the scores corresponding to nodes of
the phylogenetic tree (\code{$projected$phylo_scores}, which can then be
used to compute the node shapes using \code{\link{extract_shapes}}. The
position of these shapes in morphospace is estimated using
\code{\link[mvMORPH]{mvgls}} which allows specification of different
phenotypic evolutionary models, passed through the \code{evmodel} argument
(with the exception of the default model, Brownian motion, estimated
using \code{\link[phytools]{fastAnc}}).
}
\examples{
#load and extract relevant data, packages and information
library(magrittr)
data("tails")
shapes <- tails$shapes
species <- tails$data$species
sp_shapes <- expected_shapes(shapes, species)
tree <- tails$tree
links <- tails$links

#generate basic morphospace, add sampled shapes, species mean shapes, and
#phylogenetic structure
mspace(shapes, links = links, mag = 0.7, axes = c(1,2), cex.ldm = 0) \%>\%
  proj_shapes(shapes = shapes, col = c(1:13)[species], pch = 1,
              cex = 0.7) \%>\%
  proj_phylogeny(shapes = sp_shapes, tree = tree)
}
\references{
Clavel, J., Escarguel, G., & Merceron, G. (2015). \emph{mvMORPH: an R package
for fitting multivariate evolutionary models to morphometric data}.
Methods in Ecology and Evolution, 6(11), 1311-1319.
Revell, L.J. (2012). \emph{phytools: an R package for phylogenetic
comparative biology (and other things)}. Methods in Ecology and Evolution,
3, 217-223.
}
\seealso{
\code{\link[mvMORPH]{mvgls}}, \code{\link[phytools]{fastAnc}}
}
