% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_builders.R
\name{rev_eigen}
\alias{rev_eigen}
\title{Reverse eigenanalysis-based ordination}
\usage{
rev_eigen(scores, vectors, center)
}
\arguments{
\item{scores}{A matrix of scores resulting from an ordination.}

\item{vectors}{A matrix of eigenvector coefficients used to build the
ordination.}

\item{center}{A vector with the mean values of the original variables used in
the ordination.}
}
\value{

}
\description{
Back-calculate variables in their original scale from existing
ordination data. Used internally.
}
