% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{rotate_fcoef}
\alias{rotate_fcoef}
\title{Rotate Fourier shape 180 degrees}
\usage{
rotate_fcoef(fcoef)
}
\arguments{
\item{fcoef}{The set of Fourier coefficients measuring the shape(s) to be
rotated.}
}
\value{
A set of Fourier coefficients describing the rotated outline(s).
}
\description{
Correct 180-degrees spurious rotation in closed outline shapes.
Used internally.
}
\examples{
#load shells data, plot the first shape
data("shells")
shape1 <- shells$shapes$coe[1,]
plot(inv_efourier(shape1, 300))

#rotate and plot again
shape1_rot <- rotate_fcoef(shape1)
plot(inv_efourier(shape1_rot, 300))
}
\references{
Iwata, H., & Ukai, Y. (2002). \emph{SHAPE: a computer program package for
quantitative evaluation of biological shapes based on elliptic Fourier
descriptors}. Journal of Heredity, 93(5), 384-385.
}
