% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_builders.R
\name{sample_ax}
\alias{sample_ax}
\title{Sample shapes along a morphometric axis}
\usage{
sample_ax(scores, vector, center, p, k)
}
\arguments{
\item{k}{}
}
\value{

}
\description{
Sample shapes along a morphometric axis
}
