% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{set_rotation3d}
\alias{set_rotation3d}
\title{Orient 3D shapes and get rotation matrix}
\usage{
set_rotation3d(shape, template = NULL, links = NULL, ...)
}
\arguments{
\item{shape}{An matrix containing the (x,y,z) values of a 3D landmark and/or
semilandmark configuration.}

\item{template}{An optional \code{"mesh3d"} object containing the geometry of
the structure the 3D landmarks were placed on.}

\item{links}{A list with the indices of the coordinates defining the
(following the format used in \code{Morpho}).}

\item{...}{Further arguments passed to  \code{\link[rgl:plot3d]{rgl::plot3d()}}}
}
\value{
A rotation matrix.
}
\description{
Interactively rotate 3D landmark configuration and/or meshes to
a desired orientation.
}
\details{
This function can be used to manually set a fixed orientation for 3D
background shape models depicted by \code{mspace} (argument
\code{rot.models}), to avoid manual rotation at each \code{mspace}
instance.
}
\examples{
# load packages
library(geomorph)
library(Morpho)
library(morphospace)

# load data
data("shells3D")
shapes <- shells3D$shapes
mspec_mesh <- shells3D$mesh_meanspec

# get shape of the specimen closest to the mean shape
shape <- shapes[,,geomorph::findMeanSpec(shapes)]

if (interactive()) {
  # set rotation matrix using a landmark configuration
  rotmat1 <- set_rotation3d(shape)

  # plot morphospace
  mspace(shapes, template = template, rot.models = rotmat1)

  # set rotation matrix using a landmark configuration
  rotmat2 <- set_rotation3d(mspec_mesh)

  # plot morphospace
  mspace(shapes, template = template, rot.models = rotmat2)
}
}
\seealso{
\code{\link{plot_morphogrid3d}}, \code{\link{mspace}}
}
