% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{shapes_mat}
\alias{shapes_mat}
\title{Identify and arrange shape descriptors}
\usage{
shapes_mat(shapes)
}
\arguments{
\item{shapes}{Shape data.}
}
\value{
A list of length 2 containing:
\itemize{
\item \code{$datype:} {the type of geometric morphometrics data.}
\item \code{$data2d:} {the shape descriptors arranged in 2-margins matrix
format.}
}
}
\description{
Identify data type (Fourier coefficients or Procrustes shape
coordinates) and arrange them in 2-margin matrix format, suitable to be
used as input for multivariate analysis. Used internally.
}
\examples{
#apply on shells dataset
data("shells")
dat1 <- shapes_mat(shells$shapes)

#inspect results
head(dat1$data2d)
dat1$datype

#aply on tails dataset
data("tails")
dat2 <- shapes_mat(tails$shapes)

#inspect results
head(dat2$data2d)
dat2$datype
}
\keyword{internal}
