% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{shells}
\alias{shells}
\title{\emph{Ptychomya} shell outlines data set}
\format{
A list containing:
\describe{
\item{$shapes:}{ an \code{"OutCoe"} object containing a \code{137 x 28}
matrix with the normalized Fourier coefficients describing the outline
shape of the sample of fossil shells.}
\item{$sizes:}{ a vector of length 137 containing the centroid size of each
fossil shell.}
\item{$data: }{ a 4-column data frame with information about the taxonomic
classification, ammonoid biozone, estimated geochronologic age (in million
years), and geographic provenance of each fossil shell.}
}
}
\usage{
shells
}
\description{
Sample of intra- and interspecific shell shapes, quantified
as closed outlines using 7 harmonics and elliptic Fourier analysis, from
4 species of the extinct genus \emph{Ptychomya} (Bivalvia: Crassatellidae).
}
\details{
The \code{shells} data contain data from 137 specimens belonging to
4 species of the extinct bivalve genus \emph{Ptychomya}, tracking their
morphological changes through a 5 million years interval from the Lower
Cretaceous of Argentina (approximately 140 million years ago). The data
set includes the information about the extrernal shell shape (measured using
Fourier coefficients from 7 harmonics), centroid size, age (both relative,
taken from ammonoid biozones, and absolute, estimated using a combination
of the former, absolute dates, and stratigraphic data), geographic provenance,
and taxonomic classification of each fossil specimen.
}
\references{
Milla Carmona, P. S., Lazo, D. G., & Soto, I. M. (2018). Morphological evolution
of the bivalve Ptychomya through the Lower Cretaceous of Argentina. Paleobiology,
44(1), 101-117.
}
\keyword{datasets}
