% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stack}
\alias{stack}
\title{Pile shapes}
\usage{
stack(shapes, links = NULL, mshape = TRUE, ...)
}
\arguments{
\item{shapes}{A \code{k x p x n} array of superimposed landmarks.}

\item{links}{An optional list containing the indices of the coordinates
defining the wireframe in Morpho format.}

\item{mshape}{Logical; whether to plot the mean configuration.}

\item{...}{Additional arguments passed to \code{\link[=plot]{plot()}}.}
}
\value{

}
\description{
Pile shapes
}
