% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{svd_block}
\alias{svd_block}
\title{Singular value decomposition for 2 blocks of variables}
\usage{
svd_block(x, y, tree = NULL)
}
\arguments{
\item{x}{First block of variables}

\item{y}{Second block of variables}

\item{tree}{An optional \code{"phy"} object containing a phylogenetic
tree whose tip.labels match rownames of \code{x} and \code{y}.}
}
\value{
Mimics the \code{\link[=svd]{svd()}} output.
}
\description{
Just a wrapper for \code{\link[=svd]{svd()}} that returns an adequate output
when used for blocks of variables. Can deal with phylogenetic data too
using \code{ape} and \code{phytools} functions internally.
}
\examples{
#laod data and packages
library(geomorph)

data("tails")
shapes <- tails$shapes
sizes <- tails$sizes
species <- tails$data$species
tree <- tails$tree
sp_shapes <- consensus(shapes, species)[,,tree$tip.label]
sp_sizes <- cbind(tapply(sizes, species, mean))[tree$tip.label,]

#perform partial svd
svd_block(x = sizes, y = two.d.array(shapes))

#perform partial svd on phylogenetic structure
svd_block(x = sp_sizes, y = two.d.array(sp_shapes), tree = tree)
}
