% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{wings}
\alias{wings}
\title{\emph{Drosophila} wing shape data set}
\format{
A list containing:
\describe{
\item{$shapes:}{ a \code{9 x 2 x 263} array with the superimposed configurations
of landmarks marking veins intersections on the right wing of each specimen.}
\item{$sizes:}{ a vector of length 263 containing the centroid size of
each specimen's wing.}
\item{$data: }{ a 3-column data frame with information about the taxonomic
classification, host cactus, sex, line and replica of each specimen.}
\item{$links: }{ a list indicating how to connect the landmarks to create a
wireframe.}
\item{$template: }{ a 2-column matrix containing the coordinates of the
9 landmarks, followed by the coordinates defining 9 curves curve describing
wing morphology, to be warped using TPS interpolation.}
}
}
\usage{
wings
}
\description{
Sample of intra- and interspecific wing shapes, quntified
using 9 landmarks, from two species of cactophilic \emph{Drosophila}
(Diptera: Drosophilidae).
}
\details{
This data set includes wing shape data from 263 experimentally bred specimens
belonging to two recently diverged cactophilic \emph{Drosophila} species, \emph{D. koepferae}
and \emph{D. buzzattii} from northwestern Argentina. These two species show preference to
different host cacti (columnar cacti from the genus \emph{Trichocereus} and prickly pears
from the genus \emph{Opuntia}, respectively) and are considered morphologically cryptic
(usually, they can be recognized only from genital morphology). Accompanying data
include the centroid size of each wing, as well as information on the host cacti each
specimen was bred, its species, sex, isofemale line, and replica. Also included is a
template containing a series of curves describing wing outline and veins (created using
\code{build_template2d}), which will be warped using TPS interpolation.
}
\references{
Soto, E. M., Goenaga, J., Hurtado, J. P., & Hasson, E. (2012). \emph{Oviposition
and performance in natural hosts in cactophilic} Drosophila. Evolutionary Ecology,
26(4), 975-990.
}
\keyword{datasets}
