% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{adapt_model}
\alias{adapt_model}
\title{Adapt format of different linear model functions}
\usage{
adapt_model(model)
}
\arguments{
\item{model}{An object of class \code{\link[stats]{"mlm"}},
\code{\link[geomorph]{"procD.lm"}}, \code{\link[RRPP]{"lm.rrpp"}},
\code{\link[mvMORPH]{"mvgls"}} or \code{\link[mvMORPH]{"mvols"}}.}
}
\value{
A list containing the response and explanatory variables,
coefficients, original centroid of response variables (or phylogenetic
mean in the case of phylogenetic linear models), and method used for
linear model fitting.
}
\description{
Extract and format results from \code{\link[stats:lm]{stats::lm()}},
\code{\link[geomorph:procD.lm]{geomorph::procD.lm()}}, \code{\link[geomorph:procD.pgls]{geomorph::procD.pgls()}}, \code{\link[RRPP:lm.rrpp]{RRPP::lm.rrpp()}},
\code{\link[mvMORPH:mvgls]{mvMORPH::mvgls()}} and \code{\link[mvMORPH:mvols]{mvMORPH::mvols()}} for their use in the mspace
workflow and shape operations. Used internally.
}
\keyword{internal}
