% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{adapt_model}
\alias{adapt_model}
\title{Adapt format of different linear model functions}
\usage{
adapt_model(model)
}
\arguments{
\item{model}{An object of class \code{"mlm"}, \code{"procD.lm"},
\code{"lm.rrpp"}, \code{"mvgls"} or \code{"mvols"}.}
}
\value{
A list containing the response and explanatory variables,
coefficients, original centroid of response variables (or phylogenetic
mean in the case of phylogenetic linear models), fitted values, and method
used for linear model fitting.
}
\description{
Extract and format results from \code{\link[stats]{lm}},
\code{\link[geomorph]{procD.lm}}, \code{\link[geomorph]{procD.pgls}},
\code{\link[RRPP]{lm.rrpp}}, \code{\link[mvMORPH]{mvols}} and
\code{\link[mvMORPH]{mvgls}} for their use in the mspace workflow and shape
operations. Used internally.
}
\examples{
#load data
data("tails")

#perform PGLS
gmdf <- geomorph::geomorph.data.frame(shapes = expected_shapes(tails$shapes, tails$data$species),
                                      sizes = cbind(tapply(tails$sizes, tails$data$species, mean)),
                                      phy = tails$tree)

#adapt PGLS format
mod <- geomorph::procD.pgls(shapes ~ sizes, phy = phy, data = gmdf)
mod_adapted <- adapt_model(mod)

#compare contents
names(mod)
names(mod_adapted)
}
\keyword{internal}
