% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_operations.R
\name{ax_transformation}
\alias{ax_transformation}
\title{Compute shapes at the extremes of a morphometric axis}
\usage{
ax_transformation(obj, axis = 1, mag = 1)
}
\arguments{
\item{obj}{An object of class \code{"mlm"}, \code{"procD.lm"},
\code{"lm.rrpp"}, \code{"mvgls"} or \code{"mvols"}, containing a linear
model fitting a single explanatory variable to shape data, or an object of
class \code{"prcomp"}, \code{"bg_prcomp"}, \code{"pls_shapes"},
\code{"phy_pls_shapes"}, \code{"burnaby"}, \code{"phy_burnaby"},
\code{"gm.prcomp"}, \code{"pls"}, \code{"bgPCA"}, \code{"pls2B"},
\code{"phyl.pca"}, \code{"mvgls.pca"} or \code{"PCA"} with the results of
multivariate ordination of shape data.}

\item{axis}{An optional integer value specifying the axis of the multivariate
ordination which is to be represented.}

\item{mag}{Numeric; magnifying factor for representing shape transformation.}
}
\value{
A 2-margins matrix, of dimensions \code{2 x (k x p)} for the case of
landmark data and \code{2 x (4 x nb.h)} for the case of Fourier data (where
\code{nb.h} is the number of harmonics used in elliptic Fourier analysis).
}
\description{
This function computes the theoretical shapes corresponding to
the extremes of a morphometric axis, which can be supplied as an object
containing a linear model or a multivariate ordination.
}
\details{
If an object of class \code{"mlm"} fitting shape to a factor (only
two levels allowed) is supplied in \code{obj}, magnification of the axis
range is attained through bgPCA.
}
\examples{
#load tail data and packages
library(geomorph)
data("tails")
shapes <- tails$shapes
links <- tails$links

#perform PCA, compute and plot extreme shapes of PC1 at its natural range
pca <- prcomp(two.d.array(shapes))
extshapes2d <- ax_transformation(obj = pca, axis = 1, mag = 1)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes((extshapes), links = links, mshape = FALSE)

#compute and plot extreme shapes of PC2 at its natural range
extshapes2d <- ax_transformation(obj = pca, axis = 2, mag = 1)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes(extshapes, links = links, mshape = FALSE)

#compute and plot extreme shapes of PC2 magnified x2
extshapes2d <- ax_transformation(obj = pca, axis = 2, mag = 2)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes at its natural
#range
model <- lm(two.d.array(tails$shapes) ~ log(tails$sizes))
extshapes2d <- ax_transformation(obj = model, mag = 1)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes at its natural
#range
model <- lm(two.d.array(tails$shapes) ~ tails$data$sex)
extshapes2d <- ax_transformation(obj = model, mag = 1)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes magnified x2
model <- lm(two.d.array(tails$shapes) ~ tails$data$sex)
extshapes2d <- ax_transformation(obj = model, mag = 2)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes(extshapes, links = links, mshape = FALSE)
}
\references{
MacLeod, N. (2009). \emph{Form & shape models}. Palaeontological Association
Newsletter, 72(620), 14-27.
}
\seealso{
\code{\link{expected_shapes}}, \code{\link{rev_eigen}},
\code{\link[stats]{lm}}, \code{\link[geomorph]{procD.lm}},
\code{\link[geomorph]{procD.pgls}}, \code{\link[RRPP]{lm.rrpp}},
\code{\link[mvMORPH]{mvols}}, \code{\link[mvMORPH]{mvgls}},
\code{\link[geomorph]{gm.prcomp}}, \code{\link[geomorph]{two.b.pls}},
\code{\link[Morpho]{groupPCA}}, \code{\link[Morpho]{pls2B}},
\code{\link[phytools]{phyl.pca}}, \code{\link[mvMORPH]{mvgls.pca}},
\code{\link[Momocs]{PCA}}
}
