% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_operations.R
\name{ax_transformation}
\alias{ax_transformation}
\title{Compute shapes at the extremes of a morphometric axis}
\usage{
ax_transformation(obj, axis = 1, mag = 1)
}
\arguments{
\item{obj}{An object containing either a multivariate ordination of class
\code{"prcomp", "bg_prcomp", "phy_prcomp"} or \code{"pls_shape"} or a
\code{"mlm"} object fitted using \code{\link[=lm]{lm()}}.}

\item{axis}{An optional numeric value specifying the axis of the multivariate
ordination which is to be represented.}

\item{mag}{Numeric; magnifying factor for representing shape transformation.}
}
\value{
A 2-margins matrix, of dimensions \code{2 x (k x p)} for the case of
landmark data and \code{2 x (4 x nb.h)} for the case of Fourer data (where
\code{nb.h} is the number of harmonics used in elliptic Fourier analysis).
}
\description{
This function computes the theoretical shapes corresponding to
the extremes of a morphometric axis, which can be supplied as an object
containing a linear model or a multivariate ordination analysis.
}
\details{
If an object of class \code{mlm} fitting shape to a factor (only two
levels allowed) is supplied in \code{obj}, magnification of the axis range
is attained through bgPCA. Axes computed from linear models and from
supervised multivariate ordinations using the same supervising variable will
usually differ in length (i.e. shape transformations will be magnified or
attenuated) due to the assumption of the former that the explanatory
variable has been measured without error.
}
\examples{
#load tail data
data("tails")

#perform PCA, compute and plot extreme shapes of PC1 at its natural range
pca <- prcomp(geomorph::two.d.array(tails$shapes))
extshapes <- ax_transformation(obj = pca, axis = 1, mag = 1)
pile_shapes(extshapes, links = links, mshape = FALSE)

#compute and plot extreme shapes of PC2 at its natural range
extshapes <- ax_transformation(obj = pca, axis = 2, mag = 1)
pile_shapes(extshapes, links = links, mshape = FALSE)

#compute and plot extreme shapes of PC2 magnified x2
extshapes <- ax_transformation(obj = pca, axis = 2, mag = 2)
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes at its natural range
model <- lm(geomorph::two.d.array(tails$shapes) ~ log(tails$sizes))
extshapes <- ax_transformation(obj = model, mag = 1)
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes at its natural range
model <- lm(geomorph::two.d.array(tails$shapes) ~ tails$data$sex)
extshapes <- ax_transformation(obj = model, mag = 1)
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes magnified x2
model <- lm(geomorph::two.d.array(tails$shapes) ~ tails$data$sex)
extshapes <- ax_transformation(obj = model, mag = 2)
pile_shapes(extshapes, links = links, mshape = FALSE)
}
\references{
MacLeod, N. (2009). \emph{Form & shape models}. Palaeontological
Association Newsletter, 72(620), 14-27.
}
\seealso{
\code{\link{expected_shapes}}, \code{\link{rev_eigen}}
}
