% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_operations.R
\name{ax_transformation}
\alias{ax_transformation}
\title{Compute shapes at the extremes of a morphometric axis}
\usage{
ax_transformation(obj, axis = 1, mag = 1)
}
\arguments{
\item{obj}{An object of class \code{\link[stats]{"mlm"}},
\code{\link[geomorph]{"procD.lm"}}, \code{\link[RRPP]{"lm.rrpp"}},
\code{\link[mvMORPH]{"mvgls"}} or \code{\link[mvMORPH]{"mvols"}},
containing a linear model fit to shape data, or an object of class
\code{\link[stats]{"prcomp"}}, \code{\link{"bg_prcomp"}},
\code{\link{"pls_shapes"}}, \code{\link{"phy_pls_shapes"}},
\code{\link{"burnaby"}}, \code{\link{"phy_burnaby"}},
\code{\link[geomorph]{"gm.prcomp"}}, \code{\link[Morpho]{"bgPCA"}},
\code{\link[Morpho]{"pls2B"}}, \code{\link[phytools]{"phyl.pca"}} or
\code{\link[mvMORPH]{"mvgls.pca"}} with the results of multivariate
ordination of shape data.}

\item{axis}{An optional integer value specifying the axis of the multivariate
ordination which is to be represented.}

\item{mag}{Numeric; magnifying factor for representing shape transformation.}
}
\value{
A 2-margins matrix, of dimensions \code{2 x (k x p)} for the case of
landmark data and \code{2 x (4 x nb.h)} for the case of Fourier data (where
\code{nb.h} is the number of harmonics used in elliptic Fourier analysis).
}
\description{
This function computes the theoretical shapes corresponding to
the extremes of a morphometric axis, which can be supplied as an object
containing a linear model or a multivariate ordination.
}
\details{
If an object of class \code{"mlm"} fitting shape to a factor (only
two levels allowed) is supplied in \code{obj}, magnification of the axis
range is attained through bgPCA.
}
\examples{
#load tail data and packages
library(geomorph)
data("tails")
shapes <- tails$shapes
links <- tails$links

#perform PCA, compute and plot extreme shapes of PC1 at its natural range
pca <- prcomp(two.d.array(shapes))
extshapes2d <- ax_transformation(obj = pca, axis = 1, mag = 1)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes((extshapes), links = links, mshape = FALSE)

#compute and plot extreme shapes of PC2 at its natural range
extshapes2d <- ax_transformation(obj = pca, axis = 2, mag = 1)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes(extshapes, links = links, mshape = FALSE)

#compute and plot extreme shapes of PC2 magnified x2
extshapes2d <- ax_transformation(obj = pca, axis = 2, mag = 2)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes at its natural
#range
model <- lm(two.d.array(tails$shapes) ~ log(tails$sizes))
extshapes2d <- ax_transformation(obj = model, mag = 1)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes at its natural
#range
model <- lm(two.d.array(tails$shapes) ~ tails$data$sex)
extshapes2d <- ax_transformation(obj = model, mag = 1)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes(extshapes, links = links, mshape = FALSE)

#perform lm of shape on size, compute and plot extreme shapes magnified x2
model <- lm(two.d.array(tails$shapes) ~ tails$data$sex)
extshapes2d <- ax_transformation(obj = model, mag = 2)
extshapes <- arrayspecs(extshapes2d, k = ncol(shapes), p = nrow(shapes))
pile_shapes(extshapes, links = links, mshape = FALSE)
}
\references{
MacLeod, N. (2009). \emph{Form & shape models}. Palaeontological Association
Newsletter, 72(620), 14-27.
}
\seealso{
\code{\link{expected_shapes}}, \code{\link{rev_eigen}}
}
