% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{exp_var}
\alias{exp_var}
\title{Calculate percentages of variation accounted by synthetic axes}
\usage{
exp_var(ord)
}
\arguments{
\item{ord}{An ordination (i.e. a \code{"prcomp"}, \code{"bg_prcomp"},
\code{"phy_prcomp"} or \code{"pls"} object).}
}
\value{

}
\description{
Calculate the percentage of total original variation accounted
by syntethtic axes generated by different multivariate ordination methods.
}
