% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{exp_var}
\alias{exp_var}
\title{Calculate percentages of variation accounted by ordination axes}
\usage{
exp_var(ordination)
}
\arguments{
\item{ordination}{An ordination (i.e. a \code{"prcomp"}, \code{"bg_prcomp"},
\code{"phy_prcomp"}, \code{"pls_shapes"} or \code{"phy_pls_shapes"}
object).}
}
\value{
A table informing the percentages and cumulative percentages of
original variation accounted by each ordination axis.
}
\description{
Calculate the percentage of total original variation accounted
by axes generated by different multivariate ordination methods.
}
\examples{
#load data
data("shells")

#extract shapes, log sizes and species classification
species <- shells$data$species
shapes <- shells$shapes$coe
sizes <- log(shells$sizes)

#perform PCA, bgPCA, and PLS
pca <- prcomp(shapes)
bgpca <- bg_prcomp(shapes, groups = species)
pls <- pls_shapes(shapes, X = sizes)

#compare percentages of total variation accounted by the first axes when
#the analysis is unsupervised (PCA), or supervised by species (bgPCA) or
#by log size (PLS)
head(exp_var(pca))
exp_var(bgpca)
exp_var(pls)
}
\seealso{
\code{\link[stats]{prcomp}}, \code{\link{bg_prcomp}},
\code{\link{phy_prcomp}}, \code{\link{pls_shapes}}
}
