% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hulls_by_group_2D}
\alias{hulls_by_group_2D}
\title{Plot 2D convex hulls for a series of groups}
\usage{
hulls_by_group_2D(xy, fac, col = 1:nlevels(fac), ...)
}
\arguments{
\item{xy}{Coordinates of the scatterplot.}

\item{fac}{A factor grouping data points.}

\item{col}{A vector (either character or numeric) indicating the colors used
for each group.}

\item{...}{Further arguments passed to \code{\link[=polygon]{polygon()}}.}
}
\description{
Plot convex hulls for different groups in 2D scatterplots
created using the generic \code{\link[=plot]{plot()}} function.
}
\examples{
#load landmark data and nencessary packages
library(geomorph)
data("tails")
shapes <- tails$shapes
species <- tails$data$species

#perform PCA
pca <- prcomp(two.d.array(shapes))

#plot and add convex hulls
plot(pca$x)
hulls_by_group_2D(pca$x, fac = species, col = "black")
}
