% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{hulls_by_group_2D}
\alias{hulls_by_group_2D}
\title{Plot 2D convex hulls for a series of groups}
\usage{
hulls_by_group_2D(
  xy,
  fac,
  col = seq_len(nlevels(fac)),
  lty = 1,
  alpha = 0,
  ...
)
}
\arguments{
\item{xy}{Coordinates of the scatterplot.}

\item{fac}{A factor grouping data points.}

\item{col}{A vector (either character or numeric) indicating the colors used
for each group.}

\item{lty}{A numeric vector indicating the type of line used to draw hulls.}

\item{alpha}{Numeric; transparency factor for hulls.}

\item{...}{Further arguments passed to \code{\link[graphics:polygon]{graphics::polygon()}}.}
}
\description{
Plot convex hulls for different groups in 2D scatterplots
created using the generic \code{\link[graphics:plot.default]{graphics::plot()}} function. Used internally
(mostly).
}
\examples{
#load landmark data and necessary packages
library(geomorph)
data("tails")
shapes <- tails$shapes
species <- tails$data$species

#perform PCA
pca <- prcomp(two.d.array(shapes))

#plot and add convex hulls
plot(pca$x)
hulls_by_group_2D(pca$x, fac = species, col = "black")
}
\seealso{
\code{\link{ellipses_by_group_2D}}, \code{\link{hulls_by_group_3D}}
}
\keyword{internal}
