% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hulls_by_group_3D}
\alias{hulls_by_group_3D}
\title{Plot 3D convex hulls for a series of groups}
\usage{
hulls_by_group_3D(xyz, fac, col = seq_len(nlevels(fac)), ...)
}
\arguments{
\item{xyz}{Coordinates for the scatterplot}

\item{fac}{A factor grouping data points.}

\item{col}{A vector (either character or numeric) indicating the colors used
for each group.}

\item{...}{Further arguments passed to \code{\link[rgl:primitives]{rgl::triangles3d()}} (e.g.
\code{specular}, \code{alpha}).}
}
\description{
Plot convex hulls for different groups in 3D scatterplots
created using \code{rgl}. Used internally (mostly).
}
\examples{
#load landmark data and necessary packages
library(geomorph)
data("tails")
shapes <- tails$shapes
species <- tails$data$species

#perform PCA
pca <- prcomp(two.d.array(shapes))

#plot and add convex hulls
\dontrun{

plot3d(pca$x)
hulls_by_group_3D(pca$x, fac = species, col = "gray")

}
}
\keyword{internal}
