% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{inv_efourier}
\alias{inv_efourier}
\title{Inverse Fourier transform}
\usage{
inv_efourier(coe, nb.pts = 120)
}
\arguments{
\item{coe}{A vector with Fourier coefficients.}

\item{nb.pts}{Numeric, specifying the number of coordinates for sampling the
outlines.}
}
\value{
A \code{nb.pts x 2} matrix of (x,y) cartesian coordinates defining
single outline shape.
}
\description{
A wrapper for \code{\link[Momocs:efourier_i]{Momocs::efourier_i()}} to transform a
set of Fourier coefficients into (x,y) coordinates. Used internally.
}
\examples{
#load data and extract the first outline
data("shells")
shape_coe <- shells$shapes$coe[1,]

#get and plot (x,y) coordinates for using different number of coordinates
shape_xy <- inv_efourier(shape_coe, nb.pts = 10)
plot(shape_xy)
shape_xy <- inv_efourier(shape_coe, nb.pts = 30)
plot(shape_xy)
shape_xy <- inv_efourier(shape_coe, nb.pts = 100)
plot(shape_xy)
}
\seealso{
\code{\link[Momocs]{efourier_i}}
}
\keyword{internal}
