% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_builders.R
\name{morphogrid}
\alias{morphogrid}
\title{Background shape models}
\usage{
morphogrid(
  ordination,
  axes,
  datype,
  x = NULL,
  y = NULL,
  p = NULL,
  k = NULL,
  nh = 5,
  nv = 4,
  mag = 1,
  asp = NA,
  xlim = NULL,
  ylim = NULL,
  rot.models = 0,
  size.models = 1,
  asp.models = 1
)
}
\arguments{
\item{ordination}{An ordination (i.e. a \code{"prcomp"}, \code{"bg_prcomp"},
\code{"phy_prcomp"} or \code{"pls"} object).}

\item{axes}{Numeric of length 1 or 2, indicating the morphometric axes to be
plotted. If values for either \code{x} or \code{y} are provided, only the
first value of this argument is considered.}

\item{datype}{type of shape data (either "fcoef" or "landm").}

\item{x}{Optional vector with a non-morphometric variable to be plotted in
the x axis.}

\item{y}{Optional vector with a non-morphometric variable to be plotted in
the y axis.}

\item{p}{Numeric, indicating the number of landmarks/semilandmarks used (for
landmark data only).}

\item{k}{Numeric, indicating the number of cartesian dimensions of
landmarks/semilandmarks (for landmark data only).}

\item{nh}{Numeric; the number of shape models along the x axis.}

\item{nv}{Numeric; the number of shape models along the y axis.}

\item{mag}{Numeric; magnifying factor for shape models.}

\item{xlim, ylim, asp}{Standard arguments passed to the generic plot function}

\item{rot.models}{Numeric; angle (in degrees) to rotate shape models.}

\item{size.models}{Numeric; size factor for shape models.}

\item{asp.models}{Numeric; the y/x aspect ratio of shape models.}
}
\value{

}
\description{
Calculate and arrange background shape models for morphospaces.
For internal use.
}
