% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{mspace}
\alias{mspace}
\title{Create morphospace}
\usage{
mspace(
  shapes,
  axes = c(1, 2),
  links,
  p = NA,
  k = NA,
  FUN = prcomp,
  nh = 5,
  nv = 4,
  mag = 1,
  asp = NA,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  rot.models = 0,
  size.models = 1,
  asp.models = 1,
  col.model = "#708095",
  lwd.model = 1,
  points = FALSE,
  cex.ldm = 1,
  col.ldm = "black",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Create morphospace
}
