% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{mspace}
\alias{mspace}
\title{Generate morphospace}
\usage{
mspace(
  shapes = NULL,
  ord = NULL,
  axes = c(1, 2),
  links = NULL,
  template = NULL,
  FUN = stats::prcomp,
  datype = NULL,
  p = NULL,
  k = NULL,
  nh = 5,
  nv = 4,
  mag = 1,
  invax = NULL,
  asp = NA,
  rescale = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  adj_frame = c(1, 1),
  rot.models = 0,
  size.models = 1,
  asp.models = 1,
  col.models = "#708095",
  bg.models = NULL,
  lwd.models = 1,
  alpha.models = 1,
  points = FALSE,
  cex.ldm = 1,
  col.ldm = "black",
  plot = TRUE,
  models = TRUE,
  ...
)
}
\arguments{
\item{shapes}{Shape data.}

\item{ord}{Optional object of class \code{"prcomp"}, \code{"bg_prcomp"},
\code{"pls_shapes"}, \code{"phy_pls_shapes"}, \code{"burnaby"},
\code{"phy_burnaby"}, \code{"gm.prcomp"}, \code{"pls"}, \code{"bgPCA"},
\code{"pls2B"}, \code{"phyl.pca"}, \code{"mvgls.pca"} or \code{"PCA"},
containing the results of multivariate ordination of shape data. To be
used instead of the \code{shapes} argument.}

\item{axes}{Numeric of length 1 (univariate morphospace) or 2 (bivariate
morphospace), indicating the axes to be plotted.}

\item{links}{A list with the indices of the coordinates defining the
wireframe (following the format used in \code{Morpho}).}

\item{template}{Either a 2-column matrix with landmarks/semilandmarks and
template curves coordinates (for 2D shape data) or a \code{"mesh3d"}
object \strong{representing the mean shape of the sample} (for 3D shape
data). See details below.}

\item{FUN}{The function (method) to be used for ordination of shape
variation. Supported alternatives include \code{\link[stats]{prcomp}},
\code{\link{bg_prcomp}}, \code{\link{pls_shapes}}, \code{\link{burnaby}},
\code{\link[geomorph]{gm.prcomp}}, \code{\link[geomorph]{two.b.pls}},
\code{\link[Morpho]{groupPCA}}, \code{\link[phytools]{phyl.pca}} and
\code{\link[Momocs]{PCA}}. Ignored if \code{ord} is provided.}

\item{datype}{Character; the type of shape data used (either \code{"fcoef"}
or \code{"landm"}). Only required if \code{ord} is provided instead of
\code{shapes}.}

\item{p}{Numeric, indicating the number of landmarks/semilandmarks used (only
for landmark data in 2-margin matrices format).}

\item{k}{Numeric, indicating the number of cartesian dimensions of
landmarks/semilandmarks (only for landmark data in 2-margin matrices
format).}

\item{nh}{Positive integer; number of shape models along the x axis.}

\item{nv}{Positive integer; number of shape models along the y axis.}

\item{mag}{Numeric; magnifying factor for shape models.}

\item{invax}{Optional integer indicating which of the axes provided in
\code{axes} needs to be inverted (options are \code{1}, \code{2} or
\code{c(1,2)}).}

\item{rescale}{Logical; whether to re-scale background shape models so shape
variation is shown more clearly.}

\item{xlim, ylim, xlab, ylab, asp}{Standard arguments passed to the generic plot
function.}

\item{adj_frame}{Numeric of length 2, providing \emph{a posteriori} scaling
factors for the width and height of the frame, respectively.}

\item{rot.models}{Numeric; angle (in degrees) to rotate shape models.}

\item{size.models}{Numeric; size factor for shape models.}

\item{asp.models}{Numeric; the y/x aspect ratio of shape models.}

\item{col.models}{Color for wireframes/outlines.}

\item{bg.models}{Background color for outlines/meshes.}

\item{lwd.models}{Integer; width of the lines in wireframes/outlines.}

\item{alpha.models}{Numeric; transparency factor for background models
(3D only).}

\item{points}{Logical; whether to plot the scatter points.}

\item{cex.ldm}{Numeric; size of landmarks/semilandmarks in the background
models.}

\item{col.ldm}{Color of landmarks/semilandmarks in the background models.}

\item{plot}{Logical; whether to plot morphospace.}

\item{models}{Logical; whether to plot background shape models.}

\item{...}{Further arguments passed to \code{FUN}.}
}
\value{
An object of class \code{"mspace"}, which is a list containing:
\itemize{
\item{$ordination:} { a list with the output from the ordination method
used, styled in the \code{\link{prcomp}} format (typically containing at
least an \code{$x}, \code{$rotation}, and \code{$center} slots. Also
contains the type of data (\code{$datype}) and ordination method
(\code{$ordtype}) used.}
\item{$projected:} { a list containing the elements that have been
projected into the morphospace stored in the \code{"mspace"} object.
Initially includes only the background shape models
(\code{$shapemodels}), but more elements can be added using the
\code{proj_*} family of functions.}
\item{$plotinfo:} { a list containing the graphical parameters used to
create the plot. Passed to \code{\link{plot_mspace}} to regenerate
morphospaces.}
}
}
\description{
Build morphospaces using a variety of multivariate methods,
and depict shape variation represented by the resulting ordination axes.
}
\details{
This function is the heart of the \code{morphospace} workflow. It
computes a new ordination space from a sample of normalized shapes using
a variety of eigenanalysis-based multivariate methods. Alternatively,
the results of multivariate ordinations achieved using functions from
\code{geomorph}, \code{Morpho}, \code{Momocs}, \code{phytools} or
\code{mvMORPH} can be provided directly using the \code{ord} argument
(this requires specification of the \code{datype} argument, and also
of \code{p} and \code{k} for landmark data).

In total, the user can choose from a range of ordination methods to
capture the signal of interest including PCA, Between-groups PCA,
Phylogenetic PCA, 2-blocks PLS, Phylogenetically corrected 2-block PLS,
Phylogenetically-aligned component analysis, and the Burnaby approach for
orthogonalization against nuisance variables (also including a
phylogenetically corrected variant).

Have in mind that when using the \code{geomorph::two.b.pls} or
\code{Morpho::pls2B} in the context of \code{mspace}, it is assumed that
the shape variables the user wants to use to generate the morphospace are
provided in the second (aka response) block (i.e., arguments \code{A1} and
\code{X}, respectively).

\code{mspace} will generate a series of shape models depicting the range
of variation. The resulting \code{"mspace"} object stores all the
information necessary to project and/or retrieve new shapes into the
ordination space. The latter can be populated using the \code{proj_*}
family of functions and the \code{\%>\%} operator from \code{magrittr}.
Background shapes, shape axes, or other elements projected into
morphospace can be extracted for other uses through
\code{\link{extract_shapes}}.

For landmark data, representation of shape variation can be further
enhanced by providing a template, which contains additional geometric
features from the structure the landmark/semilandmarks were placed upon.
Templates will be warped using TPS interpolation to produce the set of
background shape models. For 2D landmark data, templates must be provided
as a 2-column matrix containing the actual landmarks/semilandmarks,
followed by the coordinates defining a curve or set of curves, separated
from the former and from each other by a row of \code{NA}s (see
\code{\link{build_template2d}}). For 3D landmark data, the template
must be a \code{"mesh3d"} object corresponding to the \strong{actual mean
shape of the sample} (which can be obtained using
\code{\link{expected_shapes}} + \code{\link[Morpho]{tps3d}}; see examples
below).
}
\examples{
##2D Landmark data

#load and extract relevant data and information
data("tails")
shapes <- tails$shapes
species <- tails$data$species
links <- tails$links
tree <- tails$tree
sp_shapes <- expected_shapes(shapes, species)

#generate morphospace using the basic sample of shapes, PCA as ordination
#method and the links between landmarks provided for backround models
mspace(shapes, links = links, mag = 0.7, axes = c(1,2), points = TRUE)

#increase magnification factor x2:
mspace(shapes, links = links, mag = 1.5, axes = c(1,2), points = TRUE)

#plot PCs 1 and 3
mspace(shapes, links = links, mag = 1.5, axes = c(1,3), points = TRUE)

#generate morphospace using the basic sample of shapes, bgPCA as ordination
#method and links between landmarks for backround models
mspace(shapes, links = links, FUN = bg_prcomp, groups = species, mag = 0.7,
       axes = c(1,2), invax = 1, points = TRUE)

#just create a morphospace without plotting, save into an object, and inspect
morphosp <- mspace(shapes, links = links, mag = 0.7, axes = c(1,2),
                   plot = FALSE)
morphosp #general info about the object
names(morphosp) #slots


# Generate morphospaces using functions from other packages (note that
# aside from replacing the shape argument with the ord one, additional
#information on number of dimensions and coordinates is required for
#landmark data):

#phylogenetically-aligned component analysis via geomorph::gm.prcomp
library(geomorph)
paca <- gm.prcomp(A = two.d.array(sp_shapes), phy = tree,
                  align.to.phy = TRUE)
mspace(ord = paca, datype = "landm", p = 9, k = 2, links = links,
       points = TRUE)

#phylogenetic PCA via phytools::phyl.pca
library(phytools)
phypca <- phyl.pca(Y = two.d.array(sp_shapes), tree = tree)
mspace(ord = phypca, datype = "landm", p = 9, k = 2, links = links,
       points = TRUE)

#between-groups PCA via Morpho::groupPCA
library(Morpho)
bgpca <- groupPCA(two.d.array(shapes), groups = species)
mspace(ord = bgpca, datype = "landm", p = 9, k = 2, links = links,
       points = TRUE)

#phylogenetic PCA via mvMORPH::mvgls.pca **NOTICE a $center slot containing
#the center of the original variables need to be appended to the mvgls.pca
#object; in this case, this is the phylogenetic mean
library(mvMORPH)
phypca2 <- mvgls.pca(object = mvgls(two.d.array(sp_shapes) ~ 1, tree = tree,
                     model = "BM"), plot = FALSE)
phypca2$center <- apply(two.d.array(sp_shapes), 2, phytools::fastAnc, tree = tree)[1,]
mspace(ord = phypca2, datype = "landm", p = 9, k = 2, links = links,
       points = TRUE)


#load wing data for a quick demo with templates
data("wings")
shapes <- wings$shapes
links <- wings$links
template <- wings$template

#generate morphospace using links
mspace(shapes, links = links, mag = 3, axes = c(1,2), points = TRUE)

#generate morphospace using template
mspace(shapes, template = template, mag = 3, axes = c(1,2), points = TRUE)



##3D Landmark data

\dontrun{
#load data and packages and extract relevant data and information
library(Morpho)
library(geomorph)
data("shells3D")
shapes <- shells3D$shapes
mesh_meanspec <- shells3D$mesh_meanspec

#generate morphospace. This is interactive, you need to rotate the shape by
#yourself and then press enter into the console.
mspace(shapes, mag = 1, axes = c(1,2), col.ldm = "black", cex.ldm = 2,
       points = TRUE)

#generate morphospace using a mesh template that improves visualization:
#first, get shape corresponding to shells3D$mesh_meanspec using
#geomorph::findMeanSpec
meanspec_id <- findMeanSpec(shapes)
meanspec_shape <- shapes[,,meanspec_id]

#then get the consensus shape and warp the sample mesh to get the mesh
#corresponding to the consensus using Morpho::tps3d
meanshape <- expected_shapes(shapes)
meanmesh <- tps3d(x = mesh_meanspec , refmat = meanspec_shape,
                  tarmat = meanshape)

#finally, generate morphospace providing template (this function used the
#mesh warped to the mean shape of the entire sample, hence the previous
#lines)
morphosp_3d <- mspace(shapes, mag = 1, axes = c(1,2), template = meanmesh,
                      bg.models = "gray", nh = 4, nv = 4, cex.ldm = 0,
                      points = TRUE)

#inspect the contents of the object
morphosp_3d
}


##Outline data

#load and extract relevant data and information
data("shells")
shapes <- shells$shapes$coe

#generate morphospace using all the raw variation
mspace(shapes, mag = 1, axes = c(1,2), nh = 5, nv = 4, size.models = 1,
       rescale = FALSE, asp.models = 1, bg.model = "light gray")

#shapes in the background can be rescaled to avhieve a (slightly) better
#visualization. Also, save the ordination into an object.
morphosp_F <- mspace(shapes, mag = 1, axes = c(1,2), nh = 5, nv = 4,
                     size.models = 1, asp.models = 1,
                     bg.model = "light gray")

#inspect the contents of the object
morphosp_F
}
\seealso{
\code{\link{proj_shapes}}, \code{\link{proj_groups}},
\code{\link{proj_phylogeny}}, \code{\link{proj_axis}},
\code{\link{proj_landscape}}, \code{\link{extract_shapes}},
\code{\link{prcomp}}, \code{\link{bg_prcomp}}, \code{\link{pls_shapes}},
\code{\link[geomorph]{gm.prcomp}}, \code{\link[geomorph]{two.b.pls}},
\code{\link[Morpho]{groupPCA}}, \code{\link[Morpho]{pls2B}},
\code{\link[phytools]{phyl.pca}}, \code{\link[mvMORPH]{mvgls.pca}},
\code{\link[geomorph]{gm.prcomp}}, \code{\link[Morpho]{groupPCA}},
\code{\link[Morpho]{pls2B}}, \code{\link[phytools]{phyl.pca}},
\code{\link[mvMORPH]{mvgls.pca}}, \code{\link[Momocs]{PCA}}
}
