% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{phy_pls2b}
\alias{phy_pls2b}
\title{Phylogenetic Two-blocks Partial Least Squares}
\usage{
phy_pls2b(x, y, tree, LOOCV = FALSE, recompute = FALSE)
}
\arguments{
\item{x}{A matrix with one or more variables as columns and observations as
rows, representing the first block.}

\item{y}{A matrix with one or more variables as columns and observations as
rows, representing the second block.}

\item{tree}{A \code{"phy"} object containing a phylogenetic tree. Tip labels
should match the row number and names from \code{x} and \code{y}.}

\item{LOOCV}{Logical; whether to apply leave-one-out cross-validation.}

\item{recompute}{Logical; whether to re-compute rotation matrix using the
scores resulting from LOOCV.}
}
\value{
A \code{"phy_pls2b"} object, containing:
\itemize{
\item \code{$values:} {vector of singular values accounting for the
covariation among blocks explained by each par of axes.}
\item \code{$xrotation:} {matrix of vector coefficients for the first
block.}
\item \code{$yrotation:} {matrix of vector coefficients for the second
block.}
\item \code{$xcenter:} {the mean values of the original variables from
the first block.}
\item \code{$ycenter:} {the mean values of the original variables from
the second block.}
\item \code{$xtotvar:} {the sum of the variances from the original
variables from the second block.}
\item \code{$ytotvar:} {the sum of the variances from the original
variables from the second block.}
}
}
\description{
Performs phylogenetic 2B Partial Least Squares allowing for leave-one-out cross-validation. Experimental.
}
\details{
Similar to \code{\link{pls2b}} but cases are linked by phylogenetic
relationships. Like in the case of \code{\link{phy_prcomp}}, the resulting
axes maximize the residual covariation among blocks left after removing
covariation among blocks accounted by phylogenetic history (assuming a
Brownian model of evolution an 100\% phylogenetic signal, which is
equivalent to setting \code{method = "BM"} in \code{\link[=phyl.pca]{phyl.pca()}}).

This method display the same variational properties than phylogenetic PCA,
(i.e. centering on the phylogenetic mean; orientation of scores reflect
non-phylogenetic covariation but their variance is not scaled and thus
contain phylogenetic information; for the latter reason, variance of
scores and eigenvalues differ; scores are correlated; etc; see Polly
et al. 2013).

As in other supervised multivariate ordination functions from
\code{morphospace}, leave-one-out cross-validation can be implemented to
alleviate the spurious covariation among blocks that is produced when
the number of variables exceeds the number of cases.
}
\examples{
#load data
data("tails")

#extract mean log sizes and shapes for all species, as well as the phylogenetic tree
sp_shapes <- consensus(shapes = tails$shapes, index = tails$data$species)
sp_sizes <- tapply(X = log(tails$sizes), INDEX = tails$data$species, FUN = mean)
tree <- tails$tree

#perform phylogenetic PLS
ppls <- phy_pls2b(y = geomorph::two.d.array(sp_shapes), x = cbind(sp_sizes), tree = tree)

#look at the results
names(ppls) #the contents of the resulting object
plot(ppls$xscores, ppls$yscores) #ordination


#pls_shapes achieves identical results, but it is formatted to be used more
#easily when analyzing shape variables and its output is compatible with other
#functions from morphospace
ppls2 <- pls_shapes(shapes = sp_shapes, X = cbind(sp_sizes), tree = tree)

names(ppls2) #the contents of the resulting object
exp_var(ppls2) #variance explained by each axis
plot(ppls2$x2, ppls2$x) #ordination
}
\references{
Rohlf, F. J., & Corti, M. (2000). \emph{Use of two-block partial
least-squares to study covariation in shape}. Systematic Biology, 49(4),
740-753.

Bookstein, F. L. (2019). \emph{Pathologies of between-groups principal
components analysis in geometric morphometrics}. Evolutionary Biology, 46(4),
271-302.

Polly, P. D., Lawing, A. M., Fabre, A. C., & Goswami, A. (2013).
\emph{Phylogenetic principal components analysis and geometric
morphometrics}. Hystrix, 24(1), 33.

Monteiro, L. R. (2013). \emph{Morphometrics and the comparative method:
studying the evolution of biological shape}. Hystrix, the Italian Journal
of Mammalogy, 24(1), 25-32.
}
\seealso{
\code{\link{pls_shapes}}, \code{\link{pls2b}},
\code{\link{phy_prcomp}}, \code{\link{exp_var}}
}
