% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{phy_prcomp}
\alias{phy_prcomp}
\title{Phylogenetic Principal Component Analysis}
\usage{
phy_prcomp(x, tree, corr = FALSE, ...)
}
\arguments{
\item{x}{A matrix with one or more variables as columns and observations as
rows. Row must be named and match tip labels from the phylogenetic
\code{tree}.}

\item{tree}{A \code{"phy"} object containing a phylogenetic tree. Tip labels
should match the row names from \code{x}.}

\item{corr}{Logical; whether to use correlation instead of covariance matrix
as input.}

\item{...}{}
}
\value{
A \code{"phy_prcomp"} object formatted following the \code{"prcomp"}
class:
\itemize{
\item \code{$sdev:} {the standard deviations of the principal components
(i.e. the square roots of the eigenvalues of the covariance/correlation
matrix).}
\item \code{$rotation:} {a matrix of eigenvector coefficients.}
\item \code{$center:} {the phylogenetic mean (i.e. the shape estimated
for the root of the tree).}
\item \code{$totvar:} {the sum of the variances from all the original
variables.}
\item \code{$lambda, $logL:} {fitted value of lambda and log-likelihood
of the model; see \code{\link[phytools]{phyl.pca}}.}
}
}
\description{
A wrapper for \code{\link[=pyl.pca]{pyl.pca()}} from \code{phytools}.
}
\details{
Phylogenetic PCA finds the linear combination of variables (in the
the context of \code{morphospace} will generally be a series of shapes
arranged as 2-margin matrix) maximizing the residual variation left after
removing covariation explained by phylogenetic history (i.e. they reflect
the covariance that would correspond to a star phylogeny).

Phylogenetic PCA has some important differences to regular PCA. First, the
resulting ordination is centered around the phylogenetic mean (i.e. the
values estimated for the root of the tree) instead of the overall centroid
of the original variables. More importantly, both phylogenetic PCA's
eigenvectors and eigenvalues have been constructed using covariation that
is independent of phylogenetic structure. However only the orientation of
the scores on those axes, and not their variances, reflect this adjustment.
In other words, orientation of the phylogenetic PC axes are devoid of
phylogenetic structure, but magnitudes measured in the resulting morphospace
(e.g. distances, variances) still retain phylogenetic information. Stemming
from this same fact, the variances computed using phylogenetic scores differ
from the ones calculated from the phylogenetic eigenvalues (which represent
the total amount of variance among the taxa after removing covariance
accounted by phylogeny, although their magnitude depends on the units used
to measure branch length), and do not necessarily decrease for subordinate
axes. Also, the set of phylogenetic scores are not uncorrelated, meaning
they can contain redundant information. For more details, see Polly et al.
2013.
}
\examples{
#load data
data("tails")

#compute mean shapes for all species and extract the phylogenetic tree
sp_shapes <- consensus(shapes = tails$shapes, index = tails$data$species)
tree <- tails$tree

#perform phylogenetic PCA
ppca <- phy_prcomp(x = geomorph::two.d.array(sp_shapes), tree = tree)

#look at the results
names(ppca) #the contents of the resulting object
exp_var(ppca) #variance explained by each axis
plot(ppca$x) #ordination
}
\references{
Revell, L. J. (2009). \emph{Size-correction and principal
components for interspecific comparative studies}. Evolution, 63, 3258-3268.

Polly, P. D., Lawing, A. M., Fabre, A. C., & Goswami, A. (2013).
\emph{Phylogenetic principal components analysis and geometric
morphometrics}. Hystrix, 24(1), 33.

Monteiro, L. R. (2013). \emph{Morphometrics and the comparative method:
studying the evolution of biological shape}. Hystrix, the Italian Journal
of Mammalogy, 24(1), 25-32.
}
\seealso{
\code{\link[phytools]{phyl.pca}}, \code{\link[base]{prcomp}},
\code{\link{exp_var}}
}
