% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{phy_prcomp}
\alias{phy_prcomp}
\title{Phylogenetic Principal Component Analysis}
\usage{
phy_prcomp(x, tree, corr = FALSE, ...)
}
\arguments{
\item{x}{A matrix with one or more variables as columns and observations as
rows. Row must be named and match tip labels from the phylogenetic
\code{tree}.}

\item{tree}{A \code{"phy"} object containing a phylogenetic tree. Tip labels
should match the row names from \code{x}.}

\item{corr}{Logical; whether to use correlation instead of covariance matrix
as input.}

\item{...}{}
}
\value{

}
\description{
A wrapper for \code{\link[=pyl.pca]{pyl.pca()}} from \code{phytools}.
}
\seealso{
\code{\link[phytools]{phyl.pca}}, \code{\link[base]{prcomp}}
}
