% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pile_shapes}
\alias{pile_shapes}
\title{Pile shapes}
\usage{
pile_shapes(shapes, links = NULL, mshape = TRUE, ...)
}
\arguments{
\item{shapes}{Shape data.}

\item{links}{An optional list with the indices of the coordinates defining
the wireframe (following the format used in \code{Morpho}).}

\item{mshape}{Logical; whether to plot the mean configuration.}

\item{...}{Additional arguments passed to \code{\link[graphics]{lines}} or
\code{\link[rgl]{lines3d}}.}
}
\value{
\code{Node}
}
\description{
Superimpose all the shapes in the sample.
}
\examples{
#load 2D landmark data
data("tails")
shapes <- tails$shapes
links <- tails$links

#pile shapes
pile_shapes(shapes, mshape = FALSE) #bare
pile_shapes(shapes, mshape = FALSE, links = links) #with links
pile_shapes(shapes, mshape = TRUE, links = links) #with links and mean shape


#load 3D landmark data
data("shells3D")
shapes <- shells3D$shapes

#pile shapes
pile_shapes(shapes, mshape = FALSE) #bare
pile_shapes(shapes, mshape = FALSE, links = list(1:10)) #with false links
pile_shapes(shapes, mshape = TRUE, links = list(1:10)) #false links + mshape


#load outline data
data("shells")
shapes <- shells$shapes
links <- shells$links

#pile shapes
pile_shapes(shapes, mshape = FALSE) #bare
pile_shapes(shapes, mshape = TRUE, links = links) #with mean shape
}
