% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{plot_mspace}
\alias{plot_mspace}
\title{Plot morphospaces}
\usage{
plot_mspace(
  mspace,
  axes,
  x = NULL,
  y = NULL,
  nh,
  nv,
  mag,
  points = TRUE,
  mshapes = TRUE,
  groups = TRUE,
  phylo = TRUE,
  asp,
  xlab,
  ylab,
  xlim = NULL,
  ylim = NULL,
  size.models,
  asp.models,
  rot.models,
  col.models,
  bg.models,
  lwd.models,
  cex.ldm,
  col.ldm,
  pch.points = 1,
  col.points = 1,
  cex.points = 1,
  col.groups = 1,
  pch.groups = 16,
  cex.groups = 1,
  lwd.branches = 1
)
}
\arguments{
\item{mspace}{An \code{"mspace"} object created using \code{\link[=mspace]{mspace()}}.}

\item{axes}{Numeric of length 1 or 2, indicating the axes to be plotted.}

\item{x}{Optional vector with a non-morphometric variable to be plotted in
the x axis. Alternatively, a \code{"phy"} object can be provided.}

\item{y}{Optional vector with a non-morphometric variable to be plotted in
the y axis. Alternatively, a \code{"phy"} object can be provided.}

\item{nh}{Numeric; the number of shape models along the x axis.}

\item{nv}{Numeric; the number of shape models along the y axis.}

\item{mag}{Numeric; magnifying factor for shape models.}

\item{points}{Logical; whether to plot the scatter points corresponding to
the sampled shapes stored in \code{mspace$x}.}

\item{mshapes}{Logical; whether to plot the scatter points corresponding to
groups' mean shapes stored in \code{mspace$gr_centroids}.}

\item{groups}{Logical; whether to plot the convex hulls enclosing the groups
stored in \code{mspace$gr_class}}

\item{phylo}{Logical; whether to plot the phylogenetic relationships stored
in \code{mspace$phylo}}

\item{xlim, ylim, xlab, ylab, asp}{Standard arguments passed to the generic plot
function}

\item{size.models}{Numeric; size factor for shape models.}

\item{asp.models}{Numeric; the y/x aspect ratio of shape models.}

\item{rot.models}{Numeric; angle (in degrees) to rotate shape models.}

\item{col.models}{The color for wireframes/outlines.}

\item{bg.models}{Background color for outlines.}

\item{lwd.models}{Numeric; the width of the lines in wireframes/outlines.}

\item{cex.ldm}{Numeric; size of landmarks/semilandmarks in the background
models.}

\item{col.ldm}{The color of landmarks/semilandmarks in the background models.}

\item{pch.points}{Numeric; the symbol of the scatter points corresponding to
sampled shapes.}

\item{col.points}{The color of the scatter points corresponding to sampled
shapes.}

\item{cex.points}{Numeric; the size of the scatter points corresponding to
sampled shapes.}

\item{col.groups}{The color of the scatter points corresponding to groups
mean shapes.}

\item{pch.groups}{Numeric; the symbol of the scatter points corresponding to
groups mean shapes.}

\item{cex.groups}{Numeric; the size of the scatter points corresponding to
groups mean shapes.}

\item{lwd.branches}{Numeric; the width of the lines depicting phylogenetic
branches.}
}
\value{

}
\description{
Flexible deployment of morphospaces, including their combination
with other variables or a phylogeny.
}
