% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{plot_mspace}
\alias{plot_mspace}
\title{Plot morphospaces and combine them with other elements}
\usage{
plot_mspace(
  mspace,
  axes,
  links = NULL,
  template = NULL,
  x = NULL,
  y = NULL,
  nh,
  nv,
  mag,
  invax = NULL,
  points = TRUE,
  mshapes = TRUE,
  groups = TRUE,
  phylo = TRUE,
  asp = NA,
  xlab,
  ylab,
  xlim = NULL,
  ylim = NULL,
  size.models,
  asp.models,
  rot.models,
  col.models,
  bg.models,
  lwd.models,
  alpha.models,
  cex.ldm,
  col.ldm,
  pch.points = 1,
  col.points = 1,
  cex.points = 1,
  col.groups = 1,
  pch.groups = 16,
  cex.groups = 1,
  lwd.branches = 1
)
}
\arguments{
\item{mspace}{An \code{"mspace"} object created using \code{\link[=mspace]{mspace()}}.}

\item{axes}{Numeric of length 1 or 2, indicating the axes to be plotted.}

\item{links}{A list with the indices of the coordinates defining the
wireframe (following the format used in \code{Morpho}).}

\item{template}{A 2-column matrix containing 1) the actual
landmarks/semilandmarks being analized, followed by 2) the (x,y) cartesian
coordinates defining a curve or set of curves that will be warped using the
deformation interpolated from the changes between landmarks/semilandmarks
(the actual positions of which must be marked with a row of NA, see the
tails dataset).}

\item{x}{Optional vector with a non-morphometric variable to be plotted in
the x axis. Alternatively, a \code{"phy"} object can be provided.}

\item{y}{Optional vector with a non-morphometric variable to be plotted in
the y axis. Alternatively, a \code{"phy"} object can be provided.}

\item{nh}{Numeric; the number of shape models along the x axis.}

\item{nv}{Numeric; the number of shape models along the y axis.}

\item{mag}{Numeric; magnifying factor for shape models.}

\item{invax}{Optional numeric indicating which of the axes provided in
\code{axes} needs to be inverted (optionsare \code{1}, \code{2} or
\code{c(1,2)}).}

\item{points}{Logical; whether to plot the scatter points corresponding to
the sampled shapes stored in \code{mspace$x}.}

\item{mshapes}{Logical; whether to plot the scatter points corresponding to
groups' mean shapes stored in \code{mspace$gr_centroids}.}

\item{groups}{Logical; whether to plot the convex hulls enclosing the groups
stored in \code{mspace$gr_class}}

\item{phylo}{Logical; whether to plot the phylogenetic relationships stored
in \code{mspace$phylo}}

\item{xlim, ylim, xlab, ylab, asp}{Standard arguments passed to the generic plot
function.}

\item{size.models}{Numeric; size factor for shape models.}

\item{asp.models}{Numeric; the y/x aspect ratio of shape models.}

\item{rot.models}{Numeric; angle (in degrees) to rotate shape models.}

\item{col.models}{The color for wireframes/outlines.}

\item{bg.models}{Background color for outlines.}

\item{lwd.models}{Numeric; the width of the lines in wireframes/outlines.}

\item{alpha.models}{Numeric; transparency factor for background models (3D only).}

\item{cex.ldm}{Numeric; size of landmarks/semilandmarks in the background
models.}

\item{col.ldm}{The color of landmarks/semilandmarks in the background models.}

\item{pch.points}{Numeric; the symbol of the scatter points corresponding to
sampled shapes.}

\item{col.points}{The color of the scatter points corresponding to sampled
shapes.}

\item{cex.points}{Numeric; the size of the scatter points corresponding to
sampled shapes.}

\item{col.groups}{The color of the scatter points corresponding to groups
mean shapes.}

\item{pch.groups}{Numeric; the symbol of the scatter points corresponding to
groups mean shapes.}

\item{cex.groups}{Numeric; the size of the scatter points corresponding to
groups mean shapes.}

\item{lwd.branches}{Numeric; the width of the lines depicting phylogenetic
branches.}
}
\description{
Flexible representation of morphospaces, including their combination
with other variables or a phylogeny.
}
\details{
This function allows to plot the morphospace contained in \code{mspace}
objects already in existence, either preserving the graphical attributes stamped
at during the pipeline or modifying one or more of them, so there is no need to
execute the pipeline every time morphospaces need to be visualized and/or changed.

Also, \code{plot_mspace} expands the range of graphical options available
beyond 'pure' morphospaces. If a numeric non-shape variable (assumed to have
been measured for the same specimens in \code{mspace$x}) is feeded to one of
\code{x} or \code{y}, a 'hybrid' morphospace is produced (i.e. the bivariate
plot will be constructed from the combination of \code{x} or \code{y} and a
morphometric axis; shape models in the background will represent variation
only for the latter). If instead a \code{"phy"} object (assumed to describe
the phylogenetic relationships among tips scores stored in
\code{mspace$phylo_scores}) is feeded to one of \code{x} or \code{y}, a vertical
or horizontal phenogram will be deployed (the x/y axis range will correspond to
branch lengths, so caution should be exercised when interpreting the output).
}
\examples{
#load and extract relevant data, packages and information
library(magrittr)
data("tails")
shapes <- tails$shapes
species <- tails$data$species
sizes <- tails$sizes
sp_shapes <- consensus(shapes, species)
sp_sizes <- cbind(tapply(sizes, species, mean))
tree <- tails$tree
links <- tails$links
template <- tails$template

#generate basic morphospace, add sampled shapes, species mean shapes, species
#classification, and phylogenetic structure
msp <- mspace(shapes, axes = c(1,2), plot = FALSE) \%>\%
 proj_shapes(shapes = shapes) \%>\%
 proj_consensus(shapes = sp_shapes) \%>\%
 proj_groups(groups = species) \%>\%
 proj_phylogeny(tree = tree)

##Plotting 'pure' morphospaces:

#plot mspace object as it is
plot_mspace(msp, axes = c(1,2),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#add colors for points, by species
plot_mspace(msp, axes = c(1,2), col.points = species, col.groups = 1:nlevels(species),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#add links
plot_mspace(msp, axes = c(1,2), links = links,
            col.points = species, col.groups = 1:nlevels(species),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#change number and sizes of shape models in the background
plot_mspace(msp, axes = c(1,2), nh = 2, nv = 2, links = links, size.models = 2,
            col.points = species, col.groups = 1:nlevels(species),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#magnify deformation
plot_mspace(msp, axes = c(1,2), mag = 1.5, nh = 2, nv = 2, links = links, size.models = 2,
            col.points = species, col.groups = 1:nlevels(species),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#change links for templates, highlight landmarks
plot_mspace(msp, axes = c(1,2), mag = 1.5, template = template, col.points = species,
            col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#change axes 1,2 for 1,3
plot_mspace(msp, axes = c(1,3), mag = 1.5, template = template, col.points = species,
            col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)

#represent ranges and centroids of groups
plot_mspace(msp, axes = c(1,3), mag = 1.5, template = template, col.points = species,
            col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = TRUE, mshapes = TRUE, phylo = FALSE)

#add phylogeny
plot_mspace(msp, axes = c(1,3), mag = 1.5, template = template, col.points = species,
            col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = TRUE, mshapes = TRUE, phylo = TRUE)



##Plotting 'hybrid' morphospaces:

#plot size against first PC (scaling factors got a bit weird because of the scale of sizes)
plot_mspace(msp, x = sizes,  axes = 1, links = links, asp.models = 0.0004, size.models = 6000,
            col.points = species, col.groups = 1:nlevels(species), pch.points = 16,
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE, xlab = "Centroid size")

#same but with template
plot_mspace(msp, x = sizes,  axes = 1, template = template, asp.models = 0.0004, size.models = 6000,
            col.points = species, col.groups = 1:nlevels(species), pch.points = 16, cex.ldm = 0,
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE, xlab = "Centroid size")



##Plotting phenograms:

#plot horizontal phenogram against PC1
plot_mspace(msp, x = tree,  axes = 1, links = links, asp.models = 0.5, size.models = 5,
            col.points = species, col.groups = 1:nlevels(species), pch.points = 16,
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE, xlab = "Branch lengths")

#plot horizontal phenogram against PC1
plot_mspace(msp, y = tree,  axes = 2, links = links, asp.models = 20, size.models = 1,
            col.points = species, col.groups = 1:nlevels(species), pch.points = 16,
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE, ylab = "Branch lengths")
}
