% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{pls2b}
\alias{pls2b}
\title{Two-block Partial Least Squares}
\usage{
pls2b(y, x, LOOCV = FALSE, recompute = FALSE)
}
\arguments{
\item{y}{A matrix with one or more variables as columns and observations as
rows, representing the first block.}

\item{x}{A matrix with one or more variables as columns and observations as
rows, representing the second block.}

\item{LOOCV}{Logical; whether to apply leave-one-out cross-validation.}

\item{recompute}{Logical; whether to re-compute rotation matrix using the
scores resulting from LOOCV.}
}
\value{

}
\description{
Perform Partial Least Squares allowing for leave-one-out
cross-validation, which is useful one the number of variables exceeds the
number of observations (i.e., alleviates spurious covariation between
variables).
}
