% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{pls_shapes}
\alias{pls_shapes}
\title{2B Partial Least Squares for shape data}
\usage{
pls_shapes(shapes, x, LOOCV = FALSE, recompute = FALSE)
}
\arguments{
\item{shapes}{Shape data.}

\item{x}{A matrix with variables as columns and observations as rows,
representing the external variables that supervise ordination.}

\item{LOOCV}{Logical; whether to apply leave-one-out cross-validation.}

\item{recompute}{Logical; whether to re-compute rotation matrix using the
scores resulting from LOOCV.}
}
\value{

}
\description{
A wrapper for \code{\link[=pls2b]{pls2b()}} aimed specifically at synthesizing
covariation between shape data and other external, non-shape variable(s).
}
