% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_analyses.R
\name{pls_shapes}
\alias{pls_shapes}
\title{2B Partial Least Squares for shape data}
\usage{
pls_shapes(x, shapes, tree = NULL, LOOCV = FALSE, recompute = FALSE)
}
\arguments{
\item{x}{A matrix with variables as columns and observations as rows,
representing the external variables that supervise ordination
(corresponding to the first block).}

\item{shapes}{Shape data (corresponding to the second block of
\code{pls2b}.}

\item{tree}{A \code{"phy"} object containing a phylogenetic tree. Tip labels
should match the row number and names from \code{x} and \code{y}.}

\item{LOOCV}{Logical; whether to apply leave-one-out cross-validation.}

\item{recompute}{Logical; whether to re-compute rotation matrix using the
scores resulting from LOOCV.}
}
\value{
A \code{"pls"} or \code{"phy_pls"} object formatted following the
\code{"prcomp"} class:
\itemize{
\item \code{$sdev:} {the standard deviation of the PLS axis/axis of the
shape block.}
\item \code{$rotation:} {a matrix of vector coefficients for the shape
block.}
\item \code{$center:} {the mean values of the original variables from
the shape block.}
\item \code{$totvar:} {the sum of the variances from the original
variables in the shape block.}
\item \code{$supblock:} {the scores from the supervizing block (i.e.
the \code{x} scores.}
}
\description{
A wrapper for \code{\link[=pls2b]{pls2b()}} or \code{\link[=phylo_pls2b]{phylo_pls2b()}} aimed specifically
at synthesizing covariation between shape data and other external,
non-shape variable(s).
}
\details{
This function finds the linear combination maximizing covariation
between a block of variables assumed to be shape variables formatted as
a 2-margins matrix and another block which can be either another set of
shape variables or one or more non-shape variables for supervizing the
analysis.

It has been reported that PLS (as an algebraic equivalent of bgPCA)
produces spurious covariation between blocks when the number of variables
exceeds the number of observations (which is a common situation in
geometric morphometrics analyses). This problem can be alleviated by
carrying out a leave-one-out cross-validation (LOOCV; i.e. each
observation is excluded from the calculation of PLS axes before its
projection in the resulting ordination as a way to calculate its score).
}
\references{
Rohlf, F. J., & Corti, M. (2000). \emph{Use of two-block partial
least-squares to study covariation in shape}. Systematic Biology, 49(4),
740-753.

Zelditch, M. L., Swiderski, D. L., & Sheets, H. D. (2012).
\emph{Geometric morphometrics for biologists: A primer}. 2nd ed. Academic
Press.

Bookstein, F. L. (2019). \emph{Pathologies of between-groups principal
components analysis in geometric morphometrics}. Evolutionary Biology, 46(4),
271-302.
}
\seealso{
\code{\link{pls2b}}, \code{\link{phylo_pls2b}}
}
