% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_consensus}
\alias{proj_consensus}
\title{Project consensus shape(s) into morphospace}
\usage{
proj_consensus(mspace, shapes, pipe = TRUE, ...)
}
\arguments{
\item{mspace}{An \code{"mspace"} object.}

\item{shapes}{Shapes data.}

\item{pipe}{Logical; is the function being included in a pipe?}

\item{...}{Further arguments passed to \code{\link[graphics:points]{graphics::points()}}.}
}
\value{
If a plot device with a morphospace is open, shapes feeded to
\code{shapes} are projected into morphospace. If \code{pipe = FALSE} the
corresponding scores are returned invisibly. If \code{pipe = TRUE} the
supplied \code{mspace} object will be modified by adding a new
\code{$gr_centroids} slot as well as a number of graphical parameters,
and returned invisibly.
}
\description{
Project one or more mean shapes into an existing morphospace.
}
\details{
The purpose of this function is to add the scores corresponding
to groups' mean shapes to \code{mspace} objects during pipeline. Otherwise,
it does the same than \code{proj_shapes}.
}
\examples{
#load and extract relevant data, packages and information
library(magrittr)
data("shells")
shapes <- shells$shapes
species <- shells$data$species
sp_shapes <- expected_shapes(shapes, species)

#generate basic morphospace, add sampled and consensus shapes
mspace(shapes, mag = 0.7, axes = c(1,2), bg.model = "light gray") \%>\%
  proj_shapes(shapes = shapes, col = c(1:4)[species], pch = 1, cex = 0.7) \%>\%
  proj_consensus(shapes = sp_shapes, pch = 21, bg = 1:4, cex = 2)
}
\seealso{
\code{\link{expected_shapes}}
}
