% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_consensus}
\alias{proj_consensus}
\title{Project consensus shape(s) into morphospace}
\usage{
proj_consensus(shapes, mspace, pipe = TRUE, ...)
}
\arguments{
\item{shapes}{Shapes data.}

\item{mspace}{An \code{"mspace"} object created using \code{\link[=mspace]{mspace()}}.}

\item{pipe}{Logical; is the function being included in a pipe?}

\item{...}{Further arguments passed to \code{\link[=points]{points()}}.}
}
\value{

}
\description{
Project one or more mean shapes into an existing morphospace.
}
