% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_builders.R
\name{proj_eigen}
\alias{proj_eigen}
\title{Project cases into existing eigenanalysis-based ordination}
\usage{
proj_eigen(x, vectors, center)
}
\arguments{
\item{x}{A matrix with the original variables used for the ordination as
columns, and observations as rows.}

\item{vectors}{A matrix of eigenvector coefficients used to build the
ordination.}

\item{center}{A vector with the mean values of the original variables used in
the ordination.}
}
\value{

}
\description{
Project variables in their original scale into an existing
ordination. Used internally.
}
