% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_groups}
\alias{proj_groups}
\title{Delimit groups in morphospace}
\usage{
proj_groups(mspace, groups, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Delimit groups in morphospace
}
