% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_phylogeny}
\alias{proj_phylogeny}
\title{Project phylogenetic structure into morphospace}
\usage{
proj_phylogeny(tree, mspace, pipe = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Project phylogenetic structure into morphospace
}
