% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_phylogeny}
\alias{proj_phylogeny}
\title{Project phylogenetic structure into morphospace}
\usage{
proj_phylogeny(tree, mspace, pipe = TRUE, ...)
}
\arguments{
\item{tree}{A \code{"phy"} object containing a phylogenetic tree.}

\item{mspace}{An \code{"mspace"} object created using \code{\link[=mspace]{mspace()}}.}

\item{pipe}{Logical; is the function being included in a pipe?}

\item{...}{Further arguments passed to \code{\link[=lines]{lines()}}.}
}
\value{

}
\description{
Project phylogenetic relationships among a set of shapes
(representing the consensuses of phhylogenetic terminals) into an existing
morphospace.
}
