% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_shapes}
\alias{proj_shapes}
\title{Project shapes into morphospace}
\usage{
proj_shapes(shapes, mspace, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Project shapes into morphospace
}
