% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_shapes}
\alias{proj_shapes}
\title{Project shapes into morphospace}
\usage{
proj_shapes(mspace, shapes, density = TRUE, pipe = TRUE, ...)
}
\arguments{
\item{mspace}{An \code{"mspace"} object.}

\item{shapes}{Shape data.}

\item{density}{Logical; whether to add density distribution for points
(univariate ordinations only).}

\item{pipe}{Logical; is the function being included in a pipe?}

\item{...}{Further arguments passed to \code{\link[graphics:points]{graphics::points()}}.}
}
\value{
If a plot device with a morphospace is open, shapes fed to
\code{shapes} are projected into morphospace. If \code{pipe = FALSE}
those scores are returned invisibly. If \code{pipe = TRUE} the supplied
\code{"mspace"} object will be modified by appending a \code{$scores} slot
to \code{$projected} and adding some graphical parameters (stored into the
\code{$plotinfo} slot), and returned invisibly.
}
\description{
Project a set of shapes as scatterpoints into an existing
morphospace.
}
\details{
The purpose of this function is to maintain morphospace building
and sample representation as separated steps, as well as to add flexibility
to graphical representation of scatterpoints.
}
\examples{
#load and extract relevant data, packages and information
library(magrittr)
data("wings")
shapes <- wings$shapes
cactus <- wings$data$sex
species <- wings$data$species
template <- wings$template

#generate basic morphospace, add sampled shapes
mspace(shapes, template = template, mag = 0.7, axes = c(1,2)) \%>\%
  proj_shapes(shapes = shapes)

#change colors, symbols, etc for the scatter
mspace(shapes, template = template, mag = 0.7, axes = c(1,2)) \%>\%
  proj_shapes(shapes = shapes[,,species == "Db"],  col = c("green"),
              pch = c(1,16)[cactus[species == "Db"]]) \%>\%
  proj_shapes(shapes = shapes[,,species == "Dk"],  col = c("blue"),
              pch = c(1,16)[cactus[species == "Dk"]])
}
