% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspace_workflow.R
\name{proj_shapes}
\alias{proj_shapes}
\title{Project shapes into morphospace}
\usage{
proj_shapes(shapes, mspace, pipe = TRUE, ...)
}
\arguments{
\item{shapes}{Shape data.}

\item{mspace}{An \code{"mspace"} object created using \code{\link[=mspace]{mspace()}}.}

\item{pipe}{Logical; is the function being included in a pipe?}

\item{...}{Further arguments passed to \code{\link[=points]{points()}}.}
}
\value{

}
\description{
Project a set of shapes into an existing morphospace.
}
