% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_builders.R
\name{rev_eigen}
\alias{rev_eigen}
\title{Reverse eigenanalysis-based ordination}
\usage{
rev_eigen(scores, vectors, center)
}
\arguments{
\item{center}{}
}
\value{

}
\description{
Reverse eigenanalysis-based ordination
}
