% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_builders.R
\name{shapes_mat}
\alias{shapes_mat}
\title{Identify and arrange shape descriptors}
\usage{
shapes_mat(shapes)
}
\arguments{
\item{shapes}{Shape data.}
}
\value{

}
\description{
Identify data type (Fourier coefficients or Procrustes shape
coordinates) and arrange them in matrix format, suitable to be used as
input for multivariate analysis. Used internally.
}
