% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_builders.R
\name{svd_block}
\alias{svd_block}
\title{Singular value decomposition for 2 blocks of variables}
\usage{
svd_block(x, y, tree = NULL, evmodel)
}
\arguments{
\item{x}{First block of variables.}

\item{y}{Second block of variables.}

\item{tree}{An optional \code{"phylo"} object containing a phylogenetic
tree whose tip.labels match rownames of \code{x} and \code{y}.}

\item{evmodel}{Character, specifying an evolutionary model to perform
ancestral character reconstruction; options are "BM" (Brownian motion),
"EB" (Early burst) and "lambda" (Pagel's lambda transformation) (see
\code{\link[mvMORPH]{mvgls}} for more details).}
}
\value{
Mimics the \code{\link{svd}} output.
}
\description{
Just a wrapper for \code{\link[=svd]{svd()}} that returns an adequate output
when used on blocks of variables. Can deal with phylogenetic data too
using \code{ape}, \code{phytools} and \code{mvMORPH} functions.
Used internally.
}
\examples{
#laod data and packages
library(geomorph)

data("tails")
shapes <- tails$shapes
sizes <- tails$sizes
species <- tails$data$species
tree <- tails$tree
sp_shapes <- expected_shapes(shapes, species)[,,tree$tip.label]
sp_sizes <- cbind(tapply(sizes, species, mean))[tree$tip.label,]

#perform partial svd
svd_block(x = sizes, y = two.d.array(shapes), evmodel = "BM")

#perform partial svd on phylogenetic structure
svd_block(x = sp_sizes, y = two.d.array(sp_shapes),
          tree = tree, evmodel = "BM")
}
\keyword{internal}
