% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{tails}
\alias{tails}
\title{\emph{Tyrannus} tail shape data set}
\format{
A list containing:
\describe{
\item{$shapes:}{ a \code{9 x 2 x 281} array with the symmetric component
of the superimposed configurations of landmarks marking the base of the tail,
as well as the tips of the two outermost rectrices of each specimen.}
\item{$sizes:}{ a vector of length 281 containing the centroid size of
each specimen's tail.}
\item{$data: }{ a 3-column data frame with information about the taxonomic
classification, sex and type (deep forked, DF or non-deep forked, NDF.}
\item{$tree: }{ a \code{"phy"} object containing the phylogenetic tree
of the 13 \emph{Tyrannus} species.}
\item{$links: }{ a list indicating how to connect the landmarks to create a
wireframe.}
}
}
\usage{
tails
}
\description{
Sample of intra- and interspecific tail shapes, quantified
using 9 landmarks, from the 13 species of the genus \emph{Tyrannus}
(Aves: Tyrannidae).
}
\details{
This data set contains a sample of tail shapes from the 13 species
of the genus \emph{Tyrannus}, two of which (\emph{T. savana} and
\emph{T. forficatus}) display exaggeratedly elongated tails, as well as a
considerable allometric variation and sexual dimorphism. The \code{tails}
data set  contains landmark data and centroid sizes from the tails of 281
specimens, their classification to species and sex, and the phylogenetic
relationships between \emph{Tyrannus} species (see Fasanelli et al. 2022
and references therein). To further help visualization of shapes, the links
between landmarks have also been included.

Have in mind that since the asymmetric component has been removed from
shape variation, this half the rank of a typical shape data matrix.
}
\references{
Fasanelli M. N., Milla Carmona P. S., Soto I. M., & Tuero, D.T . (2022).
\emph{Allometry, sexual selection and evolutionary lines of least
resistance shaped the evolution of exaggerated sexual traits within the
genus} Tyrannus. Journal of Evolutionary Biology, 35(5), 669 - 679.
}
\keyword{datasets}
